Network.AWS.CloudFormation

Description

AWS CloudFormation gives developers and systems administrators an easy way to create and manage a collection of related AWS resources, provisioning and updating them in an orderly and predictable fashion. You can use AWS CloudFormation’s sample templates or create your own templates to describe the AWS resources, and any associated dependencies or runtime parameters, required to run your application. You don’t need to figure out the order for provisioning AWS services or the subtleties of making those dependencies work. CloudFormation takes care of this for you. After the AWS resources are deployed, you can modify and update them in a controlled and predictable way, in effect applying version control to your AWS infrastructure the same way you do with your software.

Documentation

module Network.AWS.CloudFormation.CancelUpdateStack

module Network.AWS.CloudFormation.CreateStack

module Network.AWS.CloudFormation.DeleteStack

module Network.AWS.CloudFormation.DescribeStackEvents

module Network.AWS.CloudFormation.DescribeStackResource

module Network.AWS.CloudFormation.DescribeStackResources

module Network.AWS.CloudFormation.DescribeStacks

module Network.AWS.CloudFormation.EstimateTemplateCost

module Network.AWS.CloudFormation.GetStackPolicy

module Network.AWS.CloudFormation.GetTemplate

module Network.AWS.CloudFormation.GetTemplateSummary

module Network.AWS.CloudFormation.ListStackResources

module Network.AWS.CloudFormation.ListStacks

module Network.AWS.CloudFormation.SetStackPolicy

module Network.AWS.CloudFormation.SignalResource

module Network.AWS.CloudFormation.Types

module Network.AWS.CloudFormation.UpdateStack

module Network.AWS.CloudFormation.ValidateTemplate