Network.AWS.CloudFormation.CancelUpdateStack

Description

Cancels an update on the specified stack. If the call completes successfully, the stack will roll back the update and revert to the previous stack configuration.

Only stacks that are in the UPDATE_IN_PROGRESS state can be canceled.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_CancelUpdateStack.html

Request

data CancelUpdateStack

Request constructor

cancelUpdateStack

Arguments

:: Text
cusStackName

-> CancelUpdateStack

CancelUpdateStack constructor.

The fields accessible through corresponding lenses are:

cusStackName :: Text

Request lenses

cusStackName :: Lens' CancelUpdateStack Text

The name or the unique identifier associated with the stack.

Response

data CancelUpdateStackResponse

Response constructor

cancelUpdateStackResponse :: CancelUpdateStackResponse

CancelUpdateStackResponse constructor.