Network.AWS.CloudFormation.CreateStack

Description

Creates a stack as specified in the template. After the call completes successfully, the stack creation starts. You can check the status of the stack via the DescribeStacks API.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_CreateStack.html

Request

data CreateStack

Request constructor

createStack

Arguments

:: Text
csStackName

-> CreateStack

CreateStack constructor.

The fields accessible through corresponding lenses are:

csCapabilities :: [Capability]

csDisableRollback :: Maybe Bool

csNotificationARNs :: [Text]

csOnFailure :: Maybe OnFailure

csParameters :: [Parameter]

csStackName :: Text

csStackPolicyBody :: Maybe Text

csStackPolicyURL :: Maybe Text

csTags :: [Tag]

csTemplateBody :: Maybe Text

csTemplateURL :: Maybe Text

csTimeoutInMinutes :: Maybe Natural

Request lenses

csCapabilities :: Lens' CreateStack [Capability]

A list of capabilities that you must specify before AWS CloudFormation can create or update certain stacks. Some stack templates might include resources that can affect permissions in your AWS account. For those stacks, you must explicitly acknowledge their capabilities by specifying this parameter.

Currently, the only valid value is CAPABILITY_IAM, which is required for the following resources: AWS::CloudFormation::Stack, AWS::IAM::AccessKey, AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, AWS::IAM::Role, AWS::IAM::User, and AWS::IAM::UserToGroupAddition. If your stack template contains these resources, we recommend that you review any permissions associated with them. If you don't specify this parameter, this action returns an InsufficientCapabilities error.

csDisableRollback :: Lens' CreateStack (Maybe Bool)

Set to true to disable rollback of the stack if stack creation failed. You can specify either DisableRollback or OnFailure, but not both.

Default: false

csNotificationARNs :: Lens' CreateStack [Text]

The Simple Notification Service (SNS) topic ARNs to publish stack related events. You can find your SNS topic ARNs using the SNS console or your Command Line Interface (CLI).

csOnFailure :: Lens' CreateStack (Maybe OnFailure)

Determines what action will be taken if stack creation fails. This must be one of: DO_NOTHING, ROLLBACK, or DELETE. You can specify either OnFailure or DisableRollback, but not both.

Default: ROLLBACK

csParameters :: Lens' CreateStack [Parameter]

A list of Parameter structures that specify input parameters for the stack.

csStackName :: Lens' CreateStack Text

The name associated with the stack. The name must be unique within your AWS account.

Must contain only alphanumeric characters (case sensitive) and start with an alpha character. Maximum length of the name is 255 characters.

csStackPolicyBody :: Lens' CreateStack (Maybe Text)

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

csStackPolicyURL :: Lens' CreateStack (Maybe Text)

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

csTags :: Lens' CreateStack [Tag]

A set of user-defined Tags to associate with this stack, represented by key/value pairs. Tags defined for the stack are propagated to EC2 resources that are created as part of the stack. A maximum number of 10 tags can be specified.

csTemplateBody :: Lens' CreateStack (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

csTemplateURL :: Lens' CreateStack (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 307,200 bytes) located in an S3 bucket in the same region as the stack. For more information, go to the Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

csTimeoutInMinutes :: Lens' CreateStack (Maybe Natural)

The amount of time that can pass before the stack status becomes CREATE_FAILED; if DisableRollback is not set or is set to false, the stack will be rolled back.

Response

data CreateStackResponse

Response constructor

createStackResponse :: CreateStackResponse

CreateStackResponse constructor.

The fields accessible through corresponding lenses are:

csrStackId :: Maybe Text

Response lenses

csrStackId :: Lens' CreateStackResponse (Maybe Text)

Unique identifier of the stack.