Network.AWS.CloudFormation.CreateStackInstances

Description

Creates stack instances for the specified accounts, within the specified regions. A stack instance refers to a stack in a specific account and region. Accounts and Regions are required parameters—you must specify at least one account and one region.

Creating a Request

createStackInstances

Arguments

:: Text
csiStackSetName

-> CreateStackInstances

Creates a value of CreateStackInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csiOperationPreferences - Preferences for how AWS CloudFormation performs this stack set operation.

csiOperationId - The unique identifier for this stack set operation. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, the SDK generates one automatically. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

csiStackSetName - The name or unique ID of the stack set that you want to create stack instances from.

csiAccounts - The names of one or more AWS accounts that you want to create stack instances in the specified region(s) for.

csiRegions - The names of one or more regions where you want to create stack instances using the specified AWS account(s).

data CreateStackInstances

See: createStackInstances smart constructor.

Request Lenses

csiOperationPreferences :: Lens' CreateStackInstances (Maybe StackSetOperationPreferences)

Preferences for how AWS CloudFormation performs this stack set operation.

csiOperationId :: Lens' CreateStackInstances (Maybe Text)

The unique identifier for this stack set operation. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, the SDK generates one automatically. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

csiStackSetName :: Lens' CreateStackInstances Text

The name or unique ID of the stack set that you want to create stack instances from.

csiAccounts :: Lens' CreateStackInstances [Text]

The names of one or more AWS accounts that you want to create stack instances in the specified region(s) for.

csiRegions :: Lens' CreateStackInstances [Text]

The names of one or more regions where you want to create stack instances using the specified AWS account(s).

Destructuring the Response

createStackInstancesResponse

Arguments

:: Int
csirsResponseStatus

-> CreateStackInstancesResponse

Creates a value of CreateStackInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csirsOperationId - The unique identifier for this stack set operation.

csirsResponseStatus - -- | The response status code.

data CreateStackInstancesResponse

See: createStackInstancesResponse smart constructor.

Response Lenses

csirsOperationId :: Lens' CreateStackInstancesResponse (Maybe Text)

The unique identifier for this stack set operation.

csirsResponseStatus :: Lens' CreateStackInstancesResponse Int

- | The response status code.