Network.AWS.CloudFormation.CreateStackSet

Description

Creates a stack set.

Creating a Request

createStackSet

Arguments

:: Text
cssStackSetName

-> CreateStackSet

Creates a value of CreateStackSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cssParameters - The input parameters for the stack set template.

cssTemplateBody - The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

cssTemplateURL - The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that's located in an Amazon S3 bucket. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

cssClientRequestToken - A unique identifier for this CreateStackSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another stack set with the same name. You might retry CreateStackSet requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, the SDK generates one automatically.

cssDescription - A description of the stack set. You can use the description to identify the stack set's purpose or other important information.

cssCapabilities - A list of values that you must specify before AWS CloudFormation can create certain stack sets. Some stack set templates might include resources that can affect permissions in your AWS account—for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge their capabilities by specifying this parameter. The only valid values are CAPABILITY_IAM and CAPABILITY_NAMED_IAM. The following resources require you to specify this parameter: * AWS::IAM::AccessKey * AWS::IAM::Group * AWS::IAM::InstanceProfile * AWS::IAM::Policy * AWS::IAM::Role * AWS::IAM::User * AWS::IAM::UserToGroupAddition If your stack template contains these resources, we recommend that you review all permissions that are associated with them and edit their permissions if necessary. If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM. If you don't specify this parameter, this action returns an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

cssTags - The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. A maximum number of 50 tags can be specified. If you specify tags as part of a CreateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you don't, the entire CreateStackSet action fails with an access denied error, and the stack set is not created.

cssStackSetName - The name to associate with the stack set. The name must be unique in the region where you create your stack set.

data CreateStackSet

See: createStackSet smart constructor.

Request Lenses

cssParameters :: Lens' CreateStackSet [Parameter]

The input parameters for the stack set template.

cssTemplateBody :: Lens' CreateStackSet (Maybe Text)

The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

cssTemplateURL :: Lens' CreateStackSet (Maybe Text)

The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that's located in an Amazon S3 bucket. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

cssClientRequestToken :: Lens' CreateStackSet (Maybe Text)

A unique identifier for this CreateStackSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another stack set with the same name. You might retry CreateStackSet requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, the SDK generates one automatically.

cssDescription :: Lens' CreateStackSet (Maybe Text)

A description of the stack set. You can use the description to identify the stack set's purpose or other important information.

cssCapabilities :: Lens' CreateStackSet [Capability]

A list of values that you must specify before AWS CloudFormation can create certain stack sets. Some stack set templates might include resources that can affect permissions in your AWS account—for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge their capabilities by specifying this parameter. The only valid values are CAPABILITY_IAM and CAPABILITY_NAMED_IAM. The following resources require you to specify this parameter: * AWS::IAM::AccessKey * AWS::IAM::Group * AWS::IAM::InstanceProfile * AWS::IAM::Policy * AWS::IAM::Role * AWS::IAM::User * AWS::IAM::UserToGroupAddition If your stack template contains these resources, we recommend that you review all permissions that are associated with them and edit their permissions if necessary. If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM. If you don't specify this parameter, this action returns an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

cssTags :: Lens' CreateStackSet [Tag]

The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. A maximum number of 50 tags can be specified. If you specify tags as part of a CreateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you don't, the entire CreateStackSet action fails with an access denied error, and the stack set is not created.

cssStackSetName :: Lens' CreateStackSet Text

The name to associate with the stack set. The name must be unique in the region where you create your stack set.

Destructuring the Response

createStackSetResponse

Arguments

:: Int
cssrsResponseStatus

-> CreateStackSetResponse

Creates a value of CreateStackSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cssrsStackSetId - The ID of the stack set that you're creating.

cssrsResponseStatus - -- | The response status code.

data CreateStackSetResponse

See: createStackSetResponse smart constructor.

Response Lenses

cssrsStackSetId :: Lens' CreateStackSetResponse (Maybe Text)

The ID of the stack set that you're creating.

cssrsResponseStatus :: Lens' CreateStackSetResponse Int

- | The response status code.