Network.AWS.CloudFormation.DeleteStack

Description

Deletes a specified stack. Once the call completes successfully, stack deletion starts. Deleted stacks do not show up in the DescribeStacks API if the deletion has been completed successfully. https:/cloudformation.us-east-1.amazonaws.com ?Action=DeleteStack &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature].

Request

data DeleteStack

The input for DeleteStack action.

Request constructor

deleteStack

Arguments

:: Text
dsStackName

-> DeleteStack

Smart constructor for the minimum required parameters to construct a valid DeleteStack request.

The fields accessible through corresponding lenses are:

StackName :: Text

Request lenses

dsStackName :: Lens' DeleteStack Text

The name or the unique identifier associated with the stack.

Response

data DeleteStackResponse

Response constructor

deleteStackResponse :: DeleteStackResponse

Smart constructor for the minimum required parameters to construct a valid DeleteStackResponse response.

This constructor is provided for convenience and testing purposes.