Network.AWS.CloudFormation.DeleteStackInstances

Description

Deletes stack instances for the specified accounts, in the specified regions.

Creating a Request

deleteStackInstances

Arguments

:: Text
dsiStackSetName

-> Bool
dsiRetainStacks

-> DeleteStackInstances

Creates a value of DeleteStackInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsiOperationPreferences - Preferences for how AWS CloudFormation performs this stack set operation.

dsiOperationId - The unique identifier for this stack set operation. If you don't specify an operation ID, the SDK generates one automatically. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You can retry stack set operation requests to ensure that AWS CloudFormation successfully received them. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

dsiStackSetName - The name or unique ID of the stack set that you want to delete stack instances for.

dsiAccounts - The names of the AWS accounts that you want to delete stack instances for.

dsiRegions - The regions where you want to delete stack set instances.

dsiRetainStacks - Removes the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack or add an existing, saved stack to a new stack set.

data DeleteStackInstances

See: deleteStackInstances smart constructor.

Request Lenses

dsiOperationPreferences :: Lens' DeleteStackInstances (Maybe StackSetOperationPreferences)

Preferences for how AWS CloudFormation performs this stack set operation.

dsiOperationId :: Lens' DeleteStackInstances (Maybe Text)

The unique identifier for this stack set operation. If you don't specify an operation ID, the SDK generates one automatically. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You can retry stack set operation requests to ensure that AWS CloudFormation successfully received them. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

dsiStackSetName :: Lens' DeleteStackInstances Text

The name or unique ID of the stack set that you want to delete stack instances for.

dsiAccounts :: Lens' DeleteStackInstances [Text]

The names of the AWS accounts that you want to delete stack instances for.

dsiRegions :: Lens' DeleteStackInstances [Text]

The regions where you want to delete stack set instances.

dsiRetainStacks :: Lens' DeleteStackInstances Bool

Removes the stack instances from the specified stack set, but doesn't delete the stacks. You can't reassociate a retained stack or add an existing, saved stack to a new stack set.

Destructuring the Response

deleteStackInstancesResponse

Arguments

:: Int
dsirsResponseStatus

-> DeleteStackInstancesResponse

Creates a value of DeleteStackInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsirsOperationId - The unique identifier for this stack set operation.

dsirsResponseStatus - -- | The response status code.

data DeleteStackInstancesResponse

See: deleteStackInstancesResponse smart constructor.

Response Lenses

dsirsOperationId :: Lens' DeleteStackInstancesResponse (Maybe Text)

The unique identifier for this stack set operation.

dsirsResponseStatus :: Lens' DeleteStackInstancesResponse Int

- | The response status code.