Network.AWS.CloudFormation.DescribeChangeSet

Description

Returns the inputs for the change set and a list of changes that AWS CloudFormation will make if you execute the change set. For more information, see Updating Stacks Using Change Sets in the AWS CloudFormation User Guide.

Creating a Request

describeChangeSet

Arguments

:: Text
desChangeSetName

-> DescribeChangeSet

Creates a value of DescribeChangeSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desNextToken - A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.

desStackName - If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.

desChangeSetName - The name or Amazon Resource Name (ARN) of the change set that you want to describe.

data DescribeChangeSet

The input for the DescribeChangeSet action.

See: describeChangeSet smart constructor.

Request Lenses

desNextToken :: Lens' DescribeChangeSet (Maybe Text)

A string (provided by the DescribeChangeSet response output) that identifies the next page of information that you want to retrieve.

desStackName :: Lens' DescribeChangeSet (Maybe Text)

If you specified the name of a change set, specify the stack name or ID (ARN) of the change set you want to describe.

desChangeSetName :: Lens' DescribeChangeSet Text

The name or Amazon Resource Name (ARN) of the change set that you want to describe.

Destructuring the Response

describeChangeSetResponse

Arguments

:: Int
desrsResponseStatus

-> ChangeSetStatus
desrsStatus

-> DescribeChangeSetResponse

Creates a value of DescribeChangeSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsCreationTime - The start time when the change set was created, in UTC.

desrsChanges - A list of Change structures that describes the resources AWS CloudFormation changes if you execute the change set.

desrsNotificationARNs - The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

desrsChangeSetName - The name of the change set.

desrsExecutionStatus - If the change set execution status is AVAILABLE , you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because AWS CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

desrsChangeSetId - The ARN of the change set.

desrsNextToken - If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

desrsParameters - A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

desrsStatusReason - A description of the change set's status. For example, if your attempt to create a change set failed, AWS CloudFormation shows the error message.

desrsStackId - The ARN of the stack that is associated with the change set.

desrsDescription - Information about the change set.

desrsCapabilities - If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

desrsTags - If you execute the change set, the tags that will be associated with the stack.

desrsStackName - The name of the stack that is associated with the change set.

desrsResponseStatus - -- | The response status code.

desrsStatus - The current status of the change set, such as CREATE_IN_PROGRESS , CREATE_COMPLETE , or FAILED .

data DescribeChangeSetResponse

The output for the DescribeChangeSet action.

See: describeChangeSetResponse smart constructor.

Response Lenses

desrsCreationTime :: Lens' DescribeChangeSetResponse (Maybe UTCTime)

The start time when the change set was created, in UTC.

desrsChanges :: Lens' DescribeChangeSetResponse [Change]

A list of Change structures that describes the resources AWS CloudFormation changes if you execute the change set.

desrsNotificationARNs :: Lens' DescribeChangeSetResponse [Text]

The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

desrsChangeSetName :: Lens' DescribeChangeSetResponse (Maybe Text)

The name of the change set.

desrsExecutionStatus :: Lens' DescribeChangeSetResponse (Maybe ExecutionStatus)

If the change set execution status is AVAILABLE , you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because AWS CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

desrsChangeSetId :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN of the change set.

desrsNextToken :: Lens' DescribeChangeSetResponse (Maybe Text)

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

desrsParameters :: Lens' DescribeChangeSetResponse [Parameter]

A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

desrsStatusReason :: Lens' DescribeChangeSetResponse (Maybe Text)

A description of the change set's status. For example, if your attempt to create a change set failed, AWS CloudFormation shows the error message.

desrsStackId :: Lens' DescribeChangeSetResponse (Maybe Text)

The ARN of the stack that is associated with the change set.

desrsDescription :: Lens' DescribeChangeSetResponse (Maybe Text)

Information about the change set.

desrsCapabilities :: Lens' DescribeChangeSetResponse [Capability]

If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

desrsTags :: Lens' DescribeChangeSetResponse [Tag]

If you execute the change set, the tags that will be associated with the stack.

desrsStackName :: Lens' DescribeChangeSetResponse (Maybe Text)

The name of the stack that is associated with the change set.

desrsResponseStatus :: Lens' DescribeChangeSetResponse Int

- | The response status code.

desrsStatus :: Lens' DescribeChangeSetResponse ChangeSetStatus

The current status of the change set, such as CREATE_IN_PROGRESS , CREATE_COMPLETE , or FAILED .