Network.AWS.CloudFormation.DescribeStackEvents

Description

Returns all stack related events for a specified stack. For more information about a stack's event history, go to Stacks in the AWS CloudFormation User Guide.

You can list events for stacks that have failed to create or have been deleted by specifying the unique stack identifier (stack ID).

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeStackEvents.html

Request

data DescribeStackEvents

Request constructor

describeStackEvents :: DescribeStackEvents

DescribeStackEvents constructor.

The fields accessible through corresponding lenses are:

dseNextToken :: Maybe Text

dseStackName :: Maybe Text

Request lenses

dseNextToken :: Lens' DescribeStackEvents (Maybe Text)

String that identifies the start of the next list of events, if there is one.

Default: There is no default value.

dseStackName :: Lens' DescribeStackEvents (Maybe Text)

The name or the unique identifier associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Response

data DescribeStackEventsResponse

Response constructor

describeStackEventsResponse :: DescribeStackEventsResponse

DescribeStackEventsResponse constructor.

The fields accessible through corresponding lenses are:

dserNextToken :: Maybe Text

dserStackEvents :: [StackEvent]

Response lenses

dserNextToken :: Lens' DescribeStackEventsResponse (Maybe Text)

String that identifies the start of the next list of events, if there is one.

dserStackEvents :: Lens' DescribeStackEventsResponse [StackEvent]

A list of StackEvents structures.