Network.AWS.CloudFormation.DescribeStackResource

Description

Returns a description of the specified resource in the specified stack.

For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeStackResource.html

Request

data DescribeStackResource

Request constructor

describeStackResource

Arguments

:: Text
dsr1StackName

-> Text
dsr1LogicalResourceId

-> DescribeStackResource

DescribeStackResource constructor.

The fields accessible through corresponding lenses are:

dsr1LogicalResourceId :: Text

dsr1StackName :: Text

Request lenses

dsr1LogicalResourceId :: Lens' DescribeStackResource Text

The logical name of the resource as specified in the template.

Default: There is no default value.

dsr1StackName :: Lens' DescribeStackResource Text

The name or the unique identifier associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Response

data DescribeStackResourceResponse

Response constructor

describeStackResourceResponse :: DescribeStackResourceResponse

DescribeStackResourceResponse constructor.

The fields accessible through corresponding lenses are:

dsrrStackResourceDetail :: Maybe StackResourceDetail

Response lenses

dsrrStackResourceDetail :: Lens' DescribeStackResourceResponse (Maybe StackResourceDetail)

A StackResourceDetail structure containing the description of the specified resource in the specified stack.