Network.AWS.CloudFormation.DescribeStackSetOperation

Description

Returns the description of the specified stack set operation.

Creating a Request

describeStackSetOperation

Arguments

:: Text
dssoStackSetName

-> Text
dssoOperationId

-> DescribeStackSetOperation

Creates a value of DescribeStackSetOperation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssoStackSetName - The name or the unique stack ID of the stack set for the stack operation.

dssoOperationId - The unique ID of the stack set operation.

data DescribeStackSetOperation

See: describeStackSetOperation smart constructor.

Request Lenses

dssoStackSetName :: Lens' DescribeStackSetOperation Text

The name or the unique stack ID of the stack set for the stack operation.

dssoOperationId :: Lens' DescribeStackSetOperation Text

The unique ID of the stack set operation.

Destructuring the Response

describeStackSetOperationResponse

Arguments

:: Int
dssorsResponseStatus

-> DescribeStackSetOperationResponse

Creates a value of DescribeStackSetOperationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssorsStackSetOperation - The specified stack set operation.

dssorsResponseStatus - -- | The response status code.

data DescribeStackSetOperationResponse

See: describeStackSetOperationResponse smart constructor.

Response Lenses

dssorsStackSetOperation :: Lens' DescribeStackSetOperationResponse (Maybe StackSetOperation)

The specified stack set operation.

dssorsResponseStatus :: Lens' DescribeStackSetOperationResponse Int

- | The response status code.