Network.AWS.CloudFormation.DescribeStacks

Description

Returns the description for the specified stack; if no stack name was specified, then it returns the description for all the stacks created.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeStacks.html

Request

data DescribeStacks

Request constructor

describeStacks :: DescribeStacks

DescribeStacks constructor.

The fields accessible through corresponding lenses are:

ds1NextToken :: Maybe Text

ds1StackName :: Maybe Text

Request lenses

ds1NextToken :: Lens' DescribeStacks (Maybe Text)

String that identifies the start of the next list of stacks, if there is one.

ds1StackName :: Lens' DescribeStacks (Maybe Text)

The name or the unique identifier associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Response

data DescribeStacksResponse

Response constructor

describeStacksResponse :: DescribeStacksResponse

DescribeStacksResponse constructor.

The fields accessible through corresponding lenses are:

dsrNextToken :: Maybe Text

dsrStacks :: [Stack]

Response lenses

dsrNextToken :: Lens' DescribeStacksResponse (Maybe Text)

String that identifies the start of the next list of stacks, if there is one.

dsrStacks :: Lens' DescribeStacksResponse [Stack]

A list of stack structures.