Network.AWS.CloudFormation.EstimateTemplateCost

Description

Returns the estimated monthly cost of a template. The return value is an AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_EstimateTemplateCost.html

Request

data EstimateTemplateCost

Request constructor

estimateTemplateCost :: EstimateTemplateCost

EstimateTemplateCost constructor.

The fields accessible through corresponding lenses are:

etcParameters :: [Parameter]

etcTemplateBody :: Maybe Text

etcTemplateURL :: Maybe Text

Request lenses

etcParameters :: Lens' EstimateTemplateCost [Parameter]

A list of Parameter structures that specify input parameters.

etcTemplateBody :: Lens' EstimateTemplateCost (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.)

Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

etcTemplateURL :: Lens' EstimateTemplateCost (Maybe Text)

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

Response

data EstimateTemplateCostResponse

Response constructor

estimateTemplateCostResponse :: EstimateTemplateCostResponse

EstimateTemplateCostResponse constructor.

The fields accessible through corresponding lenses are:

etcrUrl :: Maybe Text

Response lenses

etcrUrl :: Lens' EstimateTemplateCostResponse (Maybe Text)

An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.