Network.AWS.CloudFormation.ExecuteChangeSet

Description

Updates a stack using the input information that was provided when the specified change set was created. After the call successfully completes, AWS CloudFormation starts updating the stack. Use the DescribeStacks action to view the status of the update.

When you execute a change set, AWS CloudFormation deletes all other change sets associated with the stack because they aren't valid for the updated stack.

If a stack policy is associated with the stack, AWS CloudFormation enforces the policy during the update. You can't specify a temporary stack policy that overrides the current policy.

Creating a Request

executeChangeSet

Arguments

:: Text
ecsChangeSetName

-> ExecuteChangeSet

Creates a value of ExecuteChangeSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ecsClientRequestToken - A unique identifier for this ExecuteChangeSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to execute a change set to update a stack with the same name. You might retry ExecuteChangeSet requests to ensure that AWS CloudFormation successfully received them.

ecsStackName - If you specified the name of a change set, specify the stack name or ID (ARN) that is associated with the change set you want to execute.

ecsChangeSetName - The name or ARN of the change set that you want use to update the specified stack.

data ExecuteChangeSet

The input for the ExecuteChangeSet action.

See: executeChangeSet smart constructor.

Request Lenses

ecsClientRequestToken :: Lens' ExecuteChangeSet (Maybe Text)

A unique identifier for this ExecuteChangeSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to execute a change set to update a stack with the same name. You might retry ExecuteChangeSet requests to ensure that AWS CloudFormation successfully received them.

ecsStackName :: Lens' ExecuteChangeSet (Maybe Text)

If you specified the name of a change set, specify the stack name or ID (ARN) that is associated with the change set you want to execute.

ecsChangeSetName :: Lens' ExecuteChangeSet Text

The name or ARN of the change set that you want use to update the specified stack.

Destructuring the Response

executeChangeSetResponse

Arguments

:: Int
ecsrsResponseStatus

-> ExecuteChangeSetResponse

Creates a value of ExecuteChangeSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ecsrsResponseStatus - -- | The response status code.

data ExecuteChangeSetResponse

The output for the ExecuteChangeSet action.

See: executeChangeSetResponse smart constructor.

Response Lenses

ecsrsResponseStatus :: Lens' ExecuteChangeSetResponse Int

- | The response status code.