Network.AWS.CloudFormation.GetStackPolicy

Description

Returns the stack policy for a specified stack. If a stack doesn't have a policy, a null value is returned.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_GetStackPolicy.html

Request

data GetStackPolicy

Request constructor

getStackPolicy

Arguments

:: Text
gspStackName

-> GetStackPolicy

GetStackPolicy constructor.

The fields accessible through corresponding lenses are:

gspStackName :: Text

Request lenses

gspStackName :: Lens' GetStackPolicy Text

The name or stack ID that is associated with the stack whose policy you want to get.

Response

data GetStackPolicyResponse

Response constructor

getStackPolicyResponse :: GetStackPolicyResponse

GetStackPolicyResponse constructor.

The fields accessible through corresponding lenses are:

gsprStackPolicyBody :: Maybe Text

Response lenses

gsprStackPolicyBody :: Lens' GetStackPolicyResponse (Maybe Text)

Structure containing the stack policy body. (For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide.)