Network.AWS.CloudFormation.GetTemplate

Description

Returns the template body for a specified stack. You can get the template for running or deleted stacks.

For deleted stacks, GetTemplate returns the template for up to 90 days after the stack has been deleted.

If the template does not exist, a ValidationError is returned.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_GetTemplate.html

Request

data GetTemplate

Request constructor

getTemplate

Arguments

:: Text
gtStackName

-> GetTemplate

GetTemplate constructor.

The fields accessible through corresponding lenses are:

gtStackName :: Text

Request lenses

gtStackName :: Lens' GetTemplate Text

The name or the unique identifier associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Response

data GetTemplateResponse

Response constructor

getTemplateResponse :: GetTemplateResponse

GetTemplateResponse constructor.

The fields accessible through corresponding lenses are:

gtrTemplateBody :: Maybe Text

Response lenses

gtrTemplateBody :: Lens' GetTemplateResponse (Maybe Text)

Structure containing the template body. (For more information, go to TemplateAnatomy in the AWS CloudFormation User Guide.)