Network.AWS.CloudFormation.GetTemplateSummary

Description

Returns information about a new or existing template. The GetTemplateSummary action is useful for viewing parameter information, such as default parameter values and parameter types, before you create or update a stack.

You can use the GetTemplateSummary action when you submit a template, or you can get template information for a running or deleted stack.

For deleted stacks, GetTemplateSummary returns the template information for up to 90 days after the stack has been deleted. If the template does not exist, a ValidationError is returned.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_GetTemplateSummary.html

Request

data GetTemplateSummary

Request constructor

getTemplateSummary :: GetTemplateSummary

GetTemplateSummary constructor.

The fields accessible through corresponding lenses are:

gtsStackName :: Maybe Text

gtsTemplateBody :: Maybe Text

gtsTemplateURL :: Maybe Text

Request lenses

gtsStackName :: Lens' GetTemplateSummary (Maybe Text)

The name or the unique identifier associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID.

Conditional: You must specify only one of the following parameters: StackName , TemplateBody, or TemplateURL.

gtsTemplateBody :: Lens' GetTemplateSummary (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName , TemplateBody, or TemplateURL.

gtsTemplateURL :: Lens' GetTemplateSummary (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 307,200 bytes) located in an Amazon S3 bucket. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: StackName , TemplateBody, or TemplateURL.

Response

data GetTemplateSummaryResponse

Response constructor

getTemplateSummaryResponse :: GetTemplateSummaryResponse

GetTemplateSummaryResponse constructor.

The fields accessible through corresponding lenses are:

gtsrCapabilities :: [Capability]

gtsrCapabilitiesReason :: Maybe Text

gtsrDescription :: Maybe Text

gtsrParameters :: [ParameterDeclaration]

gtsrVersion :: Maybe Text

Response lenses

gtsrCapabilities :: Lens' GetTemplateSummaryResponse [Capability]

The capabilities found within the template. Currently, AWS CloudFormation supports only the CAPABILITY_IAM capability. If your template contains IAM resources, you must specify the CAPABILITY_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

gtsrCapabilitiesReason :: Lens' GetTemplateSummaryResponse (Maybe Text)

The capabilities reason found within the template.

gtsrDescription :: Lens' GetTemplateSummaryResponse (Maybe Text)

The value that is defined in the Description property of the template.

gtsrParameters :: Lens' GetTemplateSummaryResponse [ParameterDeclaration]

A list of parameter declarations that describe various properties for each parameter.

gtsrVersion :: Lens' GetTemplateSummaryResponse (Maybe Text)

The AWS template format version, which identifies the capabilities of the template.