Network.AWS.CloudFormation.ListChangeSets

Description

Returns the ID and status of each active change set for a stack. For example, AWS CloudFormation lists change sets that are in the CREATE_IN_PROGRESS or CREATE_PENDING state.

Creating a Request

listChangeSets

Arguments

:: Text
lcsStackName

-> ListChangeSets

Creates a value of ListChangeSets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcsNextToken - A string (provided by the ListChangeSets response output) that identifies the next page of change sets that you want to retrieve.

lcsStackName - The name or the Amazon Resource Name (ARN) of the stack for which you want to list change sets.

data ListChangeSets

The input for the ListChangeSets action.

See: listChangeSets smart constructor.

Request Lenses

lcsNextToken :: Lens' ListChangeSets (Maybe Text)

A string (provided by the ListChangeSets response output) that identifies the next page of change sets that you want to retrieve.

lcsStackName :: Lens' ListChangeSets Text

The name or the Amazon Resource Name (ARN) of the stack for which you want to list change sets.

Destructuring the Response

listChangeSetsResponse

Arguments

:: Int
lcsrsResponseStatus

-> ListChangeSetsResponse

Creates a value of ListChangeSetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcsrsNextToken - If the output exceeds 1 MB, a string that identifies the next page of change sets. If there is no additional page, this value is null.

lcsrsSummaries - A list of ChangeSetSummary structures that provides the ID and status of each change set for the specified stack.

lcsrsResponseStatus - -- | The response status code.

data ListChangeSetsResponse

The output for the ListChangeSets action.

See: listChangeSetsResponse smart constructor.

Response Lenses

lcsrsNextToken :: Lens' ListChangeSetsResponse (Maybe Text)

If the output exceeds 1 MB, a string that identifies the next page of change sets. If there is no additional page, this value is null.

lcsrsSummaries :: Lens' ListChangeSetsResponse [ChangeSetSummary]

A list of ChangeSetSummary structures that provides the ID and status of each change set for the specified stack.

lcsrsResponseStatus :: Lens' ListChangeSetsResponse Int

- | The response status code.