Network.AWS.CloudFormation.ListExports

Description

Lists all exported output values in the account and region in which you call this action. Use this action to see the exported output values that you can import into other stacks. To import values, use the @Fn::ImportValue@ function.

For more information, see AWS CloudFormation Export Stack Output Values .

This operation returns paginated results.

Creating a Request

listExports :: ListExports

Creates a value of ListExports with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

leNextToken - A string (provided by the ListExports response output) that identifies the next page of exported output values that you asked to retrieve.

data ListExports

See: listExports smart constructor.

Request Lenses

leNextToken :: Lens' ListExports (Maybe Text)

A string (provided by the ListExports response output) that identifies the next page of exported output values that you asked to retrieve.

Destructuring the Response

listExportsResponse

Arguments

:: Int
lersResponseStatus

-> ListExportsResponse

Creates a value of ListExportsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lersNextToken - If the output exceeds 100 exported output values, a string that identifies the next page of exports. If there is no additional page, this value is null.

lersExports - The output for the ListExports action.

lersResponseStatus - -- | The response status code.

data ListExportsResponse

See: listExportsResponse smart constructor.

Response Lenses

lersNextToken :: Lens' ListExportsResponse (Maybe Text)

If the output exceeds 100 exported output values, a string that identifies the next page of exports. If there is no additional page, this value is null.

lersExports :: Lens' ListExportsResponse [Export]

The output for the ListExports action.

lersResponseStatus :: Lens' ListExportsResponse Int

- | The response status code.