Network.AWS.CloudFormation.ListImports

Description

Lists all stacks that are importing an exported output value. To modify or remove an exported output value, first use this action to see which stacks are using it. To see the exported output values in your account, see ListExports .

For more information about importing an exported output value, see the @Fn::ImportValue@ function.

This operation returns paginated results.

Creating a Request

listImports

Arguments

:: Text
liExportName

-> ListImports

Creates a value of ListImports with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

liNextToken - A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.

liExportName - The name of the exported output value. AWS CloudFormation returns the stack names that are importing this value.

data ListImports

See: listImports smart constructor.

Request Lenses

liNextToken :: Lens' ListImports (Maybe Text)

A string (provided by the ListImports response output) that identifies the next page of stacks that are importing the specified exported output value.

liExportName :: Lens' ListImports Text

The name of the exported output value. AWS CloudFormation returns the stack names that are importing this value.

Destructuring the Response

listImportsResponse

Arguments

:: Int
lirsResponseStatus

-> ListImportsResponse

Creates a value of ListImportsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lirsImports - A list of stack names that are importing the specified exported output value.

lirsNextToken - A string that identifies the next page of exports. If there is no additional page, this value is null.

lirsResponseStatus - -- | The response status code.

data ListImportsResponse

See: listImportsResponse smart constructor.

Response Lenses

lirsImports :: Lens' ListImportsResponse [Text]

A list of stack names that are importing the specified exported output value.

lirsNextToken :: Lens' ListImportsResponse (Maybe Text)

A string that identifies the next page of exports. If there is no additional page, this value is null.

lirsResponseStatus :: Lens' ListImportsResponse Int

- | The response status code.