Network.AWS.CloudFormation.ListStackInstances

Description

Returns summary information about stack instances that are associated with the specified stack set. You can filter for stack instances that are associated with a specific AWS account name or region.

Creating a Request

listStackInstances

Arguments

:: Text
lsiStackSetName

-> ListStackInstances

Creates a value of ListStackInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsiStackInstanceRegion - The name of the region where you want to list stack instances.

lsiNextToken - If the previous request didn't return all of the remaining results, the response's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lsiStackInstanceAccount - The name of the AWS account that you want to list stack instances for.

lsiMaxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

lsiStackSetName - The name or unique ID of the stack set that you want to list stack instances for.

data ListStackInstances

See: listStackInstances smart constructor.

Request Lenses

lsiStackInstanceRegion :: Lens' ListStackInstances (Maybe Text)

The name of the region where you want to list stack instances.

lsiNextToken :: Lens' ListStackInstances (Maybe Text)

If the previous request didn't return all of the remaining results, the response's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lsiStackInstanceAccount :: Lens' ListStackInstances (Maybe Text)

The name of the AWS account that you want to list stack instances for.

lsiMaxResults :: Lens' ListStackInstances (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

lsiStackSetName :: Lens' ListStackInstances Text

The name or unique ID of the stack set that you want to list stack instances for.

Destructuring the Response

listStackInstancesResponse

Arguments

:: Int
lsirsResponseStatus

-> ListStackInstancesResponse

Creates a value of ListStackInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsirsNextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null .

lsirsSummaries - A list of StackInstanceSummary structures that contain information about the specified stack instances.

lsirsResponseStatus - -- | The response status code.

data ListStackInstancesResponse

See: listStackInstancesResponse smart constructor.

Response Lenses

lsirsNextToken :: Lens' ListStackInstancesResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null .

lsirsSummaries :: Lens' ListStackInstancesResponse [StackInstanceSummary]

A list of StackInstanceSummary structures that contain information about the specified stack instances.

lsirsResponseStatus :: Lens' ListStackInstancesResponse Int

- | The response status code.