Network.AWS.CloudFormation.ListStackResources

Description

Returns descriptions of all resources of the specified stack. For deleted stacks, ListStackResources returns resource information for up to 90 days after the stack has been deleted. https:/cloudformation.us-east-1.amazonaws.com ?Action=ListStackResources &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2011-07-08T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] CREATE_COMPLETE DBSecurityGroup 2011-06-21T20:15:58Z gmarcteststack-dbsecuritygroup-1s5m0ez5lkk6w AWS::RDS::DBSecurityGroup CREATE_COMPLETE SampleDB 2011-06-21T20:25:57Z MyStack-sampledb-ycwhk1v830lx AWS::RDS::DBInstance CREATE_COMPLETE SampleApplication 2011-06-21T20:26:12Z MyStack-SampleApplication-1MKNASYR3RBQL AWS::ElasticBeanstalk::Application CREATE_COMPLETE SampleEnvironment 2011-06-21T20:28:48Z myst-Samp-1AGU6ERZX6M3Q AWS::ElasticBeanstalk::Environment CREATE_COMPLETE AlarmTopic 2011-06-21T20:29:06Z arn:aws:sns:us-east-1:803981987763:MyStack-AlarmTopic-SW4IQELG7RPJ AWS::SNS::Topic CREATE_COMPLETE CPUAlarmHigh 2011-06-21T20:29:23Z MyStack-CPUAlarmHigh-POBWQPDJA81F AWS::CloudWatch::Alarm 2d06e36c-ac1d-11e0-a958-f9382b6eb86b.

Request

data ListStackResources

The input for ListStackResource action.

Request constructor

mkListStackResources

Arguments

:: Text
lsrStackName

-> ListStackResources

Smart constructor for the minimum required parameters to construct a valid ListStackResources request.

The fields accessible through corresponding lenses are:

StackName :: Text

NextToken :: Maybe Text

Request lenses

lsrStackName :: Lens' ListStackResources Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

lsrNextToken :: Lens' ListStackResources (Maybe Text)

String that identifies the start of the next list of stack resource summaries, if there is one. Default: There is no default value.

Response

data ListStackResourcesResponse

The output for a ListStackResources action.

Response constructor

mkListStackResourcesResponse :: ListStackResourcesResponse

Smart constructor for the minimum required parameters to construct a valid ListStackResourcesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

StackResourceSummaries :: [StackResourceSummary]

NextToken :: Maybe Text

Response lenses

lsrrStackResourceSummaries :: Lens' ListStackResourcesResponse [StackResourceSummary]

A list of StackResourceSummary structures.

lsrrNextToken :: Lens' ListStackResourcesResponse (Maybe Text)

String that identifies the start of the next list of stack resources, if there is one.