Network.AWS.CloudFormation.ListStackSetOperationResults

Description

Returns summary information about the results of a stack set operation.

Creating a Request

listStackSetOperationResults

Arguments

:: Text
lssorStackSetName

-> Text
lssorOperationId

-> ListStackSetOperationResults

Creates a value of ListStackSetOperationResults with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lssorNextToken - If the previous request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSetOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lssorMaxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

lssorStackSetName - The name or unique ID of the stack set that you want to get operation results for.

lssorOperationId - The ID of the stack set operation.

data ListStackSetOperationResults

See: listStackSetOperationResults smart constructor.

Request Lenses

lssorNextToken :: Lens' ListStackSetOperationResults (Maybe Text)

If the previous request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSetOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lssorMaxResults :: Lens' ListStackSetOperationResults (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

lssorStackSetName :: Lens' ListStackSetOperationResults Text

The name or unique ID of the stack set that you want to get operation results for.

lssorOperationId :: Lens' ListStackSetOperationResults Text

The ID of the stack set operation.

Destructuring the Response

listStackSetOperationResultsResponse

Arguments

:: Int
lssorrsResponseStatus

-> ListStackSetOperationResultsResponse

Creates a value of ListStackSetOperationResultsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lssorrsNextToken - If the request doesn't return all results, NextToken is set to a token. To retrieve the next set of results, call ListOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, NextToken is set to null .

lssorrsSummaries - A list of StackSetOperationResultSummary structures that contain information about the specified operation results, for accounts and regions that are included in the operation.

lssorrsResponseStatus - -- | The response status code.

data ListStackSetOperationResultsResponse

See: listStackSetOperationResultsResponse smart constructor.

Response Lenses

lssorrsNextToken :: Lens' ListStackSetOperationResultsResponse (Maybe Text)

If the request doesn't return all results, NextToken is set to a token. To retrieve the next set of results, call ListOperationResults again and assign that token to the request object's NextToken parameter. If there are no remaining results, NextToken is set to null .

lssorrsSummaries :: Lens' ListStackSetOperationResultsResponse [StackSetOperationResultSummary]

A list of StackSetOperationResultSummary structures that contain information about the specified operation results, for accounts and regions that are included in the operation.

lssorrsResponseStatus :: Lens' ListStackSetOperationResultsResponse Int

- | The response status code.