Network.AWS.CloudFormation.ListStackSets

Description

Returns summary information about stack sets that are associated with the user.

Creating a Request

listStackSets :: ListStackSets

Creates a value of ListStackSets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lssStatus - The status of the stack sets that you want to get summary information about.

lssNextToken - If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSets again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lssMaxResults - The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

data ListStackSets

See: listStackSets smart constructor.

Request Lenses

lssStatus :: Lens' ListStackSets (Maybe StackSetStatus)

The status of the stack sets that you want to get summary information about.

lssNextToken :: Lens' ListStackSets (Maybe Text)

If the previous paginated request didn't return all of the remaining results, the response object's NextToken parameter value is set to a token. To retrieve the next set of results, call ListStackSets again and assign that token to the request object's NextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null .

lssMaxResults :: Lens' ListStackSets (Maybe Natural)

The maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Destructuring the Response

listStackSetsResponse

Arguments

:: Int
lssrsResponseStatus

-> ListStackSetsResponse

Creates a value of ListStackSetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lssrsNextToken - If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null .

lssrsSummaries - A list of StackSetSummary structures that contain information about the user's stack sets.

lssrsResponseStatus - -- | The response status code.

data ListStackSetsResponse

See: listStackSetsResponse smart constructor.

Response Lenses

lssrsNextToken :: Lens' ListStackSetsResponse (Maybe Text)

If the request doesn't return all of the remaining results, NextToken is set to a token. To retrieve the next set of results, call ListStackInstances again and assign that token to the request object's NextToken parameter. If the request returns all results, NextToken is set to null .

lssrsSummaries :: Lens' ListStackSetsResponse [StackSetSummary]

A list of StackSetSummary structures that contain information about the user's stack sets.

lssrsResponseStatus :: Lens' ListStackSetsResponse Int

- | The response status code.