Network.AWS.CloudFormation.SetStackPolicy

Description

Sets a stack policy for a specified stack.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_SetStackPolicy.html

Request

data SetStackPolicy

Request constructor

setStackPolicy

Arguments

:: Text
sspStackName

-> SetStackPolicy

SetStackPolicy constructor.

The fields accessible through corresponding lenses are:

sspStackName :: Text

sspStackPolicyBody :: Maybe Text

sspStackPolicyURL :: Maybe Text

Request lenses

sspStackName :: Lens' SetStackPolicy Text

The name or stack ID that you want to associate a policy with.

sspStackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

Response

data SetStackPolicyResponse

Response constructor

setStackPolicyResponse :: SetStackPolicyResponse

SetStackPolicyResponse constructor.