Network.AWS.CloudFormation.SignalResource

Description

Sends a signal to the specified resource with a success or failure status. You can use the SignalResource API in conjunction with a creation policy or update policy. AWS CloudFormation doesn't proceed with a stack creation or update until resources receive the required number of signals or the timeout period is exceeded. The SignalResource API is useful in cases where you want to send signals from anywhere other than an Amazon EC2 instance.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_SignalResource.html

Request

data SignalResource

Request constructor

signalResource

Arguments

:: Text
srStackName

-> Text
srLogicalResourceId

-> Text
srUniqueId

-> ResourceSignalStatus
srStatus

-> SignalResource

SignalResource constructor.

The fields accessible through corresponding lenses are:

srLogicalResourceId :: Text

srStackName :: Text

srStatus :: ResourceSignalStatus

srUniqueId :: Text

Request lenses

srLogicalResourceId :: Lens' SignalResource Text

The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

srStackName :: Lens' SignalResource Text

The stack name or ID that includes the resource that you want to signal.

srStatus :: Lens' SignalResource ResourceSignalStatus

The status of the signal, which is either success or failure. A failure signal causes AWS CloudFormation to immediately fail the stack creation or update.

srUniqueId :: Lens' SignalResource Text

A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

Response

data SignalResourceResponse

Response constructor

signalResourceResponse :: SignalResourceResponse

SignalResourceResponse constructor.