Network.AWS.CloudFormation.Types

Service

data CloudFormation

Version 2010-05-15 of the Amazon CloudFormation service.

Error

data RESTError :: *

XML

ns :: Text

Tag

data Tag

tag :: Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Maybe Text

tagValue :: Maybe Text

tagKey :: Lens' Tag (Maybe Text)

Required. A string used to identify this tag. You can specify a maximum of 128 characters for a tag key. Tags owned by Amazon Web Services (AWS) have the reserved prefix: 'aws:'.

tagValue :: Lens' Tag (Maybe Text)

Required. A string containing the value for this tag. You can specify a maximum of 256 characters for a tag value.

StackStatus

data StackStatus

Constructors

CreateComplete
CREATE_COMPLETE

CreateFailed
CREATE_FAILED

CreateInProgress
CREATE_IN_PROGRESS

DeleteComplete
DELETE_COMPLETE

DeleteFailed
DELETE_FAILED

DeleteInProgress
DELETE_IN_PROGRESS

RollbackComplete
ROLLBACK_COMPLETE

RollbackFailed
ROLLBACK_FAILED

RollbackInProgress
ROLLBACK_IN_PROGRESS

UpdateComplete
UPDATE_COMPLETE

UpdateCompleteCleanupInProgress
UPDATE_COMPLETE_CLEANUP_IN_PROGRESS

UpdateInProgress
UPDATE_IN_PROGRESS

UpdateRollbackComplete
UPDATE_ROLLBACK_COMPLETE

UpdateRollbackCompleteCleanupInProgress
UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS

UpdateRollbackFailed
UPDATE_ROLLBACK_FAILED

UpdateRollbackInProgress
UPDATE_ROLLBACK_IN_PROGRESS

StackEvent

data StackEvent

stackEvent

Arguments

:: Text
seStackId

-> Text
seEventId

-> Text
seStackName

-> UTCTime
seTimestamp

-> StackEvent

StackEvent constructor.

The fields accessible through corresponding lenses are:

seEventId :: Text

seLogicalResourceId :: Maybe Text

sePhysicalResourceId :: Maybe Text

seResourceProperties :: Maybe Text

seResourceStatus :: Maybe ResourceStatus

seResourceStatusReason :: Maybe Text

seResourceType :: Maybe Text

seStackId :: Text

seStackName :: Text

seTimestamp :: UTCTime

seEventId :: Lens' StackEvent Text

The unique ID of this event.

seLogicalResourceId :: Lens' StackEvent (Maybe Text)

The logical name of the resource specified in the template.

sePhysicalResourceId :: Lens' StackEvent (Maybe Text)

The name or unique identifier associated with the physical instance of the resource.

seResourceProperties :: Lens' StackEvent (Maybe Text)

BLOB of the properties used to create the resource.

seResourceStatus :: Lens' StackEvent (Maybe ResourceStatus)

Current status of the resource.

seResourceStatusReason :: Lens' StackEvent (Maybe Text)

Success/failure message associated with the resource.

seResourceType :: Lens' StackEvent (Maybe Text)

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.)

seStackId :: Lens' StackEvent Text

The unique ID name of the instance of the stack.

seStackName :: Lens' StackEvent Text

The name associated with a stack.

seTimestamp :: Lens' StackEvent UTCTime

Time the status was updated.

StackSummary

data StackSummary

stackSummary

Arguments

:: Text
ssStackName

-> UTCTime
ssCreationTime

-> StackStatus
ssStackStatus

-> StackSummary

StackSummary constructor.

The fields accessible through corresponding lenses are:

ssCreationTime :: UTCTime

ssDeletionTime :: Maybe UTCTime

ssLastUpdatedTime :: Maybe UTCTime

ssStackId :: Maybe Text

ssStackName :: Text

ssStackStatus :: StackStatus

ssStackStatusReason :: Maybe Text

ssTemplateDescription :: Maybe Text

ssCreationTime :: Lens' StackSummary UTCTime

The time the stack was created.

ssDeletionTime :: Lens' StackSummary (Maybe UTCTime)

The time the stack was deleted.

ssLastUpdatedTime :: Lens' StackSummary (Maybe UTCTime)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

ssStackId :: Lens' StackSummary (Maybe Text)

Unique stack identifier.

ssStackName :: Lens' StackSummary Text

The name associated with the stack.

ssStackStatus :: Lens' StackSummary StackStatus

The current status of the stack.

ssStackStatusReason :: Lens' StackSummary (Maybe Text)

Success/Failure message associated with the stack status.

ssTemplateDescription :: Lens' StackSummary (Maybe Text)

The template description of the template used to create the stack.

StackResourceDetail

data StackResourceDetail

stackResourceDetail

Arguments

:: Text
srdLogicalResourceId

-> Text
srdResourceType

-> UTCTime
srdLastUpdatedTimestamp

-> ResourceStatus
srdResourceStatus

-> StackResourceDetail

StackResourceDetail constructor.

The fields accessible through corresponding lenses are:

srdDescription :: Maybe Text

srdLastUpdatedTimestamp :: UTCTime

srdLogicalResourceId :: Text

srdMetadata :: Maybe Text

srdPhysicalResourceId :: Maybe Text

srdResourceStatus :: ResourceStatus

srdResourceStatusReason :: Maybe Text

srdResourceType :: Text

srdStackId :: Maybe Text

srdStackName :: Maybe Text

srdDescription :: Lens' StackResourceDetail (Maybe Text)

User defined description associated with the resource.

srdLastUpdatedTimestamp :: Lens' StackResourceDetail UTCTime

Time the status was updated.

srdLogicalResourceId :: Lens' StackResourceDetail Text

The logical name of the resource specified in the template.

srdMetadata :: Lens' StackResourceDetail (Maybe Text)

The JSON format content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the AWS CloudFormation User Guide.

srdPhysicalResourceId :: Lens' StackResourceDetail (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

srdResourceStatus :: Lens' StackResourceDetail ResourceStatus

Current status of the resource.

srdResourceStatusReason :: Lens' StackResourceDetail (Maybe Text)

Success/failure message associated with the resource.

srdResourceType :: Lens' StackResourceDetail Text

Type of resource. ((For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.)

srdStackId :: Lens' StackResourceDetail (Maybe Text)

Unique identifier of the stack.

srdStackName :: Lens' StackResourceDetail (Maybe Text)

The name associated with the stack.

ResourceStatus

data ResourceStatus

Constructors

RSCreateComplete
CREATE_COMPLETE

RSCreateFailed
CREATE_FAILED

RSCreateInProgress
CREATE_IN_PROGRESS

RSDeleteComplete
DELETE_COMPLETE

RSDeleteFailed
DELETE_FAILED

RSDeleteInProgress
DELETE_IN_PROGRESS

RSDeleteSkipped
DELETE_SKIPPED

RSUpdateComplete
UPDATE_COMPLETE

RSUpdateFailed
UPDATE_FAILED

RSUpdateInProgress
UPDATE_IN_PROGRESS

TemplateParameter

data TemplateParameter

templateParameter :: TemplateParameter

TemplateParameter constructor.

The fields accessible through corresponding lenses are:

tpDefaultValue :: Maybe Text

tpDescription :: Maybe Text

tpNoEcho :: Maybe Bool

tpParameterKey :: Maybe Text

tpDefaultValue :: Lens' TemplateParameter (Maybe Text)

The default value associated with the parameter.

tpDescription :: Lens' TemplateParameter (Maybe Text)

User defined description associated with the parameter.

tpNoEcho :: Lens' TemplateParameter (Maybe Bool)

Flag indicating whether the parameter should be displayed as plain text in logs and UIs.

tpParameterKey :: Lens' TemplateParameter (Maybe Text)

The name associated with the parameter.

ParameterDeclaration

data ParameterDeclaration

parameterDeclaration :: ParameterDeclaration

ParameterDeclaration constructor.

The fields accessible through corresponding lenses are:

pdDefaultValue :: Maybe Text

pdDescription :: Maybe Text

pdNoEcho :: Maybe Bool

pdParameterKey :: Maybe Text

pdParameterType :: Maybe Text

pdDefaultValue :: Lens' ParameterDeclaration (Maybe Text)

The default value of the parameter.

pdDescription :: Lens' ParameterDeclaration (Maybe Text)

The description that is associate with the parameter.

pdNoEcho :: Lens' ParameterDeclaration (Maybe Bool)

Flag that indicates whether the parameter value is shown as plain text in logs and in the AWS Management Console.

pdParameterKey :: Lens' ParameterDeclaration (Maybe Text)

The name that is associated with the parameter.

pdParameterType :: Lens' ParameterDeclaration (Maybe Text)

The type of parameter.

StackResource

data StackResource

stackResource

Arguments

:: Text
sr1LogicalResourceId

-> Text
sr1ResourceType

-> UTCTime
sr1Timestamp

-> ResourceStatus
sr1ResourceStatus

-> StackResource

StackResource constructor.

The fields accessible through corresponding lenses are:

sr1Description :: Maybe Text

sr1LogicalResourceId :: Text

sr1PhysicalResourceId :: Maybe Text

sr1ResourceStatus :: ResourceStatus

sr1ResourceStatusReason :: Maybe Text

sr1ResourceType :: Text

sr1StackId :: Maybe Text

sr1StackName :: Maybe Text

sr1Timestamp :: UTCTime

sr1Description :: Lens' StackResource (Maybe Text)

User defined description associated with the resource.

sr1LogicalResourceId :: Lens' StackResource Text

The logical name of the resource specified in the template.

sr1PhysicalResourceId :: Lens' StackResource (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

sr1ResourceStatus :: Lens' StackResource ResourceStatus

Current status of the resource.

sr1ResourceStatusReason :: Lens' StackResource (Maybe Text)

Success/failure message associated with the resource.

sr1ResourceType :: Lens' StackResource Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.)

sr1StackId :: Lens' StackResource (Maybe Text)

Unique identifier of the stack.

sr1StackName :: Lens' StackResource (Maybe Text)

The name associated with the stack.

sr1Timestamp :: Lens' StackResource UTCTime

Time the status was updated.

Output

data Output

output :: Output

Output constructor.

The fields accessible through corresponding lenses are:

oDescription :: Maybe Text

oOutputKey :: Maybe Text

oOutputValue :: Maybe Text

oDescription :: Lens' Output (Maybe Text)

User defined description associated with the output.

oOutputKey :: Lens' Output (Maybe Text)

The key associated with the output.

oOutputValue :: Lens' Output (Maybe Text)

The value associated with the output.

StackResourceSummary

data StackResourceSummary

stackResourceSummary

Arguments

:: Text
srsLogicalResourceId

-> Text
srsResourceType

-> UTCTime
srsLastUpdatedTimestamp

-> ResourceStatus
srsResourceStatus

-> StackResourceSummary

StackResourceSummary constructor.

The fields accessible through corresponding lenses are:

srsLastUpdatedTimestamp :: UTCTime

srsLogicalResourceId :: Text

srsPhysicalResourceId :: Maybe Text

srsResourceStatus :: ResourceStatus

srsResourceStatusReason :: Maybe Text

srsResourceType :: Text

srsLastUpdatedTimestamp :: Lens' StackResourceSummary UTCTime

Time the status was updated.

srsLogicalResourceId :: Lens' StackResourceSummary Text

The logical name of the resource specified in the template.

srsPhysicalResourceId :: Lens' StackResourceSummary (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of the resource.

srsResourceStatus :: Lens' StackResourceSummary ResourceStatus

Current status of the resource.

srsResourceStatusReason :: Lens' StackResourceSummary (Maybe Text)

Success/failure message associated with the resource.

srsResourceType :: Lens' StackResourceSummary Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.)

Capability

data Capability

Constructors

CapabilityIam
CAPABILITY_IAM

ResourceSignalStatus

data ResourceSignalStatus

Constructors

Failure
FAILURE

Success
SUCCESS

Stack

data Stack

stack

Arguments

:: Text
sStackName

-> UTCTime
sCreationTime

-> StackStatus
sStackStatus

-> Stack

Stack constructor.

The fields accessible through corresponding lenses are:

sCapabilities :: [Capability]

sCreationTime :: UTCTime

sDescription :: Maybe Text

sDisableRollback :: Maybe Bool

sLastUpdatedTime :: Maybe UTCTime

sNotificationARNs :: [Text]

sOutputs :: [Output]

sParameters :: [Parameter]

sStackId :: Maybe Text

sStackName :: Text

sStackStatus :: StackStatus

sStackStatusReason :: Maybe Text

sTags :: [Tag]

sTimeoutInMinutes :: Maybe Natural

sCapabilities :: Lens' Stack [Capability]

The capabilities allowed in the stack.

sCreationTime :: Lens' Stack UTCTime

Time at which the stack was created.

sDescription :: Lens' Stack (Maybe Text)

User defined description associated with the stack.

sDisableRollback :: Lens' Stack (Maybe Bool)

Boolean to enable or disable rollback on stack creation failures:

true: disable rollback false: enable rollback

sLastUpdatedTime :: Lens' Stack (Maybe UTCTime)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

sNotificationARNs :: Lens' Stack [Text]

SNS topic ARNs to which stack related events are published.

sOutputs :: Lens' Stack [Output]

A list of output structures.

sParameters :: Lens' Stack [Parameter]

A list of Parameter structures.

sStackId :: Lens' Stack (Maybe Text)

Unique identifier of the stack.

sStackName :: Lens' Stack Text

The name associated with the stack.

sStackStatus :: Lens' Stack StackStatus

Current status of the stack.

sStackStatusReason :: Lens' Stack (Maybe Text)

Success/failure message associated with the stack status.

sTags :: Lens' Stack [Tag]

A list of Tags that specify cost allocation information for the stack.

sTimeoutInMinutes :: Lens' Stack (Maybe Natural)

The amount of time within which stack creation should complete.

OnFailure

data OnFailure

Constructors

Delete'
DELETE

DoNothing
DO_NOTHING

Rollback
ROLLBACK

Parameter

data Parameter

parameter :: Parameter

Parameter constructor.

The fields accessible through corresponding lenses are:

pParameterKey :: Maybe Text

pParameterValue :: Maybe Text

pUsePreviousValue :: Maybe Bool

pParameterKey :: Lens' Parameter (Maybe Text)

The key associated with the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value associated with the parameter.

pUsePreviousValue :: Lens' Parameter (Maybe Bool)

During a stack update, use the existing parameter value that is being used for the stack.