Network.AWS.CloudFormation.UpdateStack

Description

Updates a stack as specified in the template. After the call completes successfully, the stack update starts. You can check the status of the stack via the DescribeStacks action.

To get a copy of the template for an existing stack, you can use the GetTemplate action.

Tags that were associated with this stack during creation time will still be associated with the stack after an UpdateStack operation.

For more information about creating an update template, updating a stack, and monitoring the progress of the update, see Updating a Stack.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_UpdateStack.html

Request

data UpdateStack

Request constructor

updateStack

Arguments

:: Text
usStackName

-> UpdateStack

UpdateStack constructor.

The fields accessible through corresponding lenses are:

usCapabilities :: [Capability]

usNotificationARNs :: [Text]

usParameters :: [Parameter]

usStackName :: Text

usStackPolicyBody :: Maybe Text

usStackPolicyDuringUpdateBody :: Maybe Text

usStackPolicyDuringUpdateURL :: Maybe Text

usStackPolicyURL :: Maybe Text

usTemplateBody :: Maybe Text

usTemplateURL :: Maybe Text

usUsePreviousTemplate :: Maybe Bool

Request lenses

usCapabilities :: Lens' UpdateStack [Capability]

A list of capabilities that you must specify before AWS CloudFormation can create or update certain stacks. Some stack templates might include resources that can affect permissions in your AWS account. For those stacks, you must explicitly acknowledge their capabilities by specifying this parameter. Currently, the only valid value is CAPABILITY_IAM, which is required for the following resources: AWS::CloudFormation::Stack, AWS::IAM::AccessKey, AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, AWS::IAM::Role, AWS::IAM::User, and AWS::IAM::UserToGroupAddition. If your stack template contains these resources, we recommend that you review any permissions associated with them. If you don't specify this parameter, this action returns an InsufficientCapabilities error.

usNotificationARNs :: Lens' UpdateStack [Text]

Update the ARNs for the Amazon SNS topics that are associated with the stack.

usParameters :: Lens' UpdateStack [Parameter]

A list of Parameter structures that specify input parameters for the stack.

usStackName :: Lens' UpdateStack Text

The name or stack ID of the stack to update.

Must contain only alphanumeric characters (case sensitive) and start with an alpha character. Maximum length of the name is 255 characters.

usStackPolicyBody :: Lens' UpdateStack (Maybe Text)

Structure containing a new stack policy body. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

usStackPolicyDuringUpdateBody :: Lens' UpdateStack (Maybe Text)

Structure containing the temporary overriding stack policy body. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

usStackPolicyDuringUpdateURL :: Lens' UpdateStack (Maybe Text)

Location of a file containing the temporary overriding stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

usStackPolicyURL :: Lens' UpdateStack (Maybe Text)

Location of a file containing the updated stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

usTemplateBody :: Lens' UpdateStack (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.)

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

usTemplateURL :: Lens' UpdateStack (Maybe Text)

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

usUsePreviousTemplate :: Lens' UpdateStack (Maybe Bool)

Reuse the existing template that is associated with the stack that you are updating.

Response

data UpdateStackResponse

Response constructor

updateStackResponse :: UpdateStackResponse

UpdateStackResponse constructor.

The fields accessible through corresponding lenses are:

usrStackId :: Maybe Text

Response lenses

usrStackId :: Lens' UpdateStackResponse (Maybe Text)

Unique identifier of the stack.