Network.AWS.CloudFormation.UpdateStackSet

Description

Updates the stack set and all associated stack instances.

Even if the stack set operation created by updating the stack set fails (completely or partially, below or above a specified failure tolerance), the stack set is updated with your changes. Subsequent CreateStackInstances calls on the specified stack set use the updated stack set.

Creating a Request

updateStackSet

Arguments

:: Text
ussStackSetName

-> UpdateStackSet

Creates a value of UpdateStackSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ussUsePreviousTemplate - Use the existing template that's associated with the stack set that you're updating. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussParameters - A list of input parameters for the stack set template.

ussOperationPreferences - Preferences for how AWS CloudFormation performs this stack set operation.

ussOperationId - The unique ID for this stack set operation. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, AWS CloudFormation generates one automatically. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

ussTemplateBody - The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussTemplateURL - The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that is located in an Amazon S3 bucket. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussDescription - A brief description of updates that you are making.

ussCapabilities - A list of values that you must specify before AWS CloudFormation can create certain stack sets. Some stack set templates might include resources that can affect permissions in your AWS account—for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge their capabilities by specifying this parameter. The only valid values are CAPABILITY_IAM and CAPABILITY_NAMED_IAM. The following resources require you to specify this parameter: * AWS::IAM::AccessKey * AWS::IAM::Group * AWS::IAM::InstanceProfile * AWS::IAM::Policy * AWS::IAM::Role * AWS::IAM::User * AWS::IAM::UserToGroupAddition If your stack template contains these resources, we recommend that you review all permissions that are associated with them and edit their permissions if necessary. If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM. If you don't specify this parameter, this action returns an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

ussTags - The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. You can specify a maximum number of 50 tags. If you specify tags for this parameter, those tags replace any list of tags that are currently associated with this stack set. This means: * If you don't specify this parameter, AWS CloudFormation doesn't modify the stack's tags. * If you specify any tags using this parameter, you must specify all the tags that you want associated with this stack set, even tags you've specifed before (for example, when creating the stack set or during a previous update of the stack set.). Any tags that you don't include in the updated list of tags are removed from the stack set, and therefore from the stacks and resources as well. * If you specify an empty value, AWS CloudFormation removes all currently associated tags. If you specify new tags as part of an UpdateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you omit tags that are currently associated with the stack set from the list of tags you specify, AWS CloudFormation assumes that you want to remove those tags from the stack set, and checks to see if you have permission to untag resources. If you don't have the necessary permission(s), the entire UpdateStackSet action fails with an access denied error, and the stack set is not updated.

ussStackSetName - The name or unique ID of the stack set that you want to update.

data UpdateStackSet

See: updateStackSet smart constructor.

Request Lenses

ussUsePreviousTemplate :: Lens' UpdateStackSet (Maybe Bool)

Use the existing template that's associated with the stack set that you're updating. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussParameters :: Lens' UpdateStackSet [Parameter]

A list of input parameters for the stack set template.

ussOperationPreferences :: Lens' UpdateStackSet (Maybe StackSetOperationPreferences)

Preferences for how AWS CloudFormation performs this stack set operation.

ussOperationId :: Lens' UpdateStackSet (Maybe Text)

The unique ID for this stack set operation. The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that AWS CloudFormation successfully received them. If you don't specify an operation ID, AWS CloudFormation generates one automatically. Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED .

ussTemplateBody :: Lens' UpdateStackSet (Maybe Text)

The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussTemplateURL :: Lens' UpdateStackSet (Maybe Text)

The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that is located in an Amazon S3 bucket. For more information, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL —or set UsePreviousTemplate to true.

ussDescription :: Lens' UpdateStackSet (Maybe Text)

A brief description of updates that you are making.

ussCapabilities :: Lens' UpdateStackSet [Capability]

A list of values that you must specify before AWS CloudFormation can create certain stack sets. Some stack set templates might include resources that can affect permissions in your AWS account—for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge their capabilities by specifying this parameter. The only valid values are CAPABILITY_IAM and CAPABILITY_NAMED_IAM. The following resources require you to specify this parameter: * AWS::IAM::AccessKey * AWS::IAM::Group * AWS::IAM::InstanceProfile * AWS::IAM::Policy * AWS::IAM::Role * AWS::IAM::User * AWS::IAM::UserToGroupAddition If your stack template contains these resources, we recommend that you review all permissions that are associated with them and edit their permissions if necessary. If you have IAM resources, you can specify either capability. If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM. If you don't specify this parameter, this action returns an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

ussTags :: Lens' UpdateStackSet [Tag]

The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. You can specify a maximum number of 50 tags. If you specify tags for this parameter, those tags replace any list of tags that are currently associated with this stack set. This means: * If you don't specify this parameter, AWS CloudFormation doesn't modify the stack's tags. * If you specify any tags using this parameter, you must specify all the tags that you want associated with this stack set, even tags you've specifed before (for example, when creating the stack set or during a previous update of the stack set.). Any tags that you don't include in the updated list of tags are removed from the stack set, and therefore from the stacks and resources as well. * If you specify an empty value, AWS CloudFormation removes all currently associated tags. If you specify new tags as part of an UpdateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you omit tags that are currently associated with the stack set from the list of tags you specify, AWS CloudFormation assumes that you want to remove those tags from the stack set, and checks to see if you have permission to untag resources. If you don't have the necessary permission(s), the entire UpdateStackSet action fails with an access denied error, and the stack set is not updated.

ussStackSetName :: Lens' UpdateStackSet Text

The name or unique ID of the stack set that you want to update.

Destructuring the Response

updateStackSetResponse

Arguments

:: Int
ussrsResponseStatus

-> UpdateStackSetResponse

Creates a value of UpdateStackSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ussrsOperationId - The unique ID for this stack set operation.

ussrsResponseStatus - -- | The response status code.

data UpdateStackSetResponse

See: updateStackSetResponse smart constructor.

Response Lenses

ussrsOperationId :: Lens' UpdateStackSetResponse (Maybe Text)

The unique ID for this stack set operation.

ussrsResponseStatus :: Lens' UpdateStackSetResponse Int

- | The response status code.