Network.AWS.CloudFormation.V2010_05_15.DeleteStack

Description

Deletes a specified stack. Once the call completes successfully, stack deletion starts. Deleted stacks do not show up in the DescribeStacks API if the deletion has been completed successfully. https:/cloudformation.us-east-1.amazonaws.com ?Action=DeleteStack &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature].

Documentation

data DeleteStack

Constructors

DeleteStack

Fields

_dsiStackName :: Text

The name or the unique identifier associated with the stack.

dsiStackName :: Iso' DeleteStack Text

data DeleteStackResponse

Constructors

DeleteStackResponse