Network.AWS.CloudFormation.V2010_05_15.DescribeStackResource

Description

Returns a description of the specified resource in the specified stack. For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted. https:/cloudformation.us-east-1.amazonaws.com ?Action=DescribeStackResource &StackName=MyStack &LogicalResourceId=MyDBInstance &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2011-07-08T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 MyStack MyDBInstance MyStack_DB1 AWS::RDS::DBInstance 2011-07-07T22:27:28Z CREATE_COMPLETE.

Documentation

data DescribeStackResource

Constructors

DescribeStackResource

Fields

_dsrjLogicalResourceId :: Text

The logical name of the resource as specified in the template. Default: There is no default value.

_dsrjStackName :: Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsrjStackName :: Lens' DescribeStackResource Text

dsrjLogicalResourceId :: Lens' DescribeStackResource Text

data DescribeStackResourceResponse

Constructors

DescribeStackResourceResponse

Fields

_dsrpStackResourceDetail :: Maybe StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

dsrpStackResourceDetail :: Iso' DescribeStackResourceResponse (Maybe StackResourceDetail)