Network.AWS.CloudFormation.V2010_05_15.EstimateTemplateCost

Description

Returns the estimated monthly cost of a template. The return value is an AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template. https:/cloudformation.us-east-1.amazonaws.com ?Action=EstimateTemplateCost &TemplateURL= https:/s3.amazonaws.comcloudformation-samples-us-east-1/Drupal_Simple.template &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2011-12-04T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] http:/calculator.s3.amazonaws.comcalc5.html?key=cf-2e351785-e821-450c-9d58-625e1e1ebfb6.

Documentation

estimateTemplateCost :: EstimateTemplateCost

Minimum specification for a EstimateTemplateCost request.

data EstimateTemplateCost

Constructors

EstimateTemplateCost

Fields

_etciParameters :: [Parameter]

A list of Parameter structures that specify input parameters.

_etciTemplateBody :: Maybe Text

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

_etciTemplateURL :: Maybe Text

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

etciTemplateURL :: Lens' EstimateTemplateCost (Maybe Text)

etciTemplateBody :: Lens' EstimateTemplateCost (Maybe Text)

etciParameters :: Lens' EstimateTemplateCost [Parameter]

data EstimateTemplateCostResponse

Constructors

EstimateTemplateCostResponse

Fields

_etcoUrl :: Maybe Text

An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

etcoUrl :: Iso' EstimateTemplateCostResponse (Maybe Text)