Network.AWS.CloudFormation.V2010_05_15.SetStackPolicy

Description

Sets a stack policy for a specified stack. https:/cloudformation.us-east-1.amazonaws.com ?Action=SetStackPolicy &StackName=MyStack &StackPolicyBody=[Stack Policy Document] &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature].

Request

data SetStackPolicy

The input for the SetStackPolicy action.

Request constructor

mkSetStackPolicy

Arguments

:: Text
sspStackName

-> SetStackPolicy

Smart constructor for the minimum required parameters to construct a valid SetStackPolicy request.

Request lenses

sspStackName :: Lens' SetStackPolicy Text

The name or stack ID that you want to associate a policy with.

sspStackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

Location of a file containing the stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

Response

data SetStackPolicyResponse