Network.AWS.CloudFormation.V2010_05_15.Types

Description

AWS CloudFormation gives developers and systems administrators an easy way to create and manage a collection of related AWS resources, provisioning and updating them in an orderly and predictable fashion. You can use AWS CloudFormation’s sample templates or create your own templates to describe the AWS resources, and any associated dependencies or runtime parameters, required to run your application. You don’t need to figure out the order for provisioning AWS services or the subtleties of making those dependencies work. CloudFormation takes care of this for you. After the AWS resources are deployed, you can modify and update them in a controlled and predictable way, in effect applying version control to your AWS infrastructure the same way you do with your software.

Service

data CloudFormation

Supported version (2010-05-15) of the AWS CloudFormation service.

Errors

data family Er a

XML

xmlOptions :: Tagged a XMLOptions

Capability

data Capability

Constructors

CapabilityCapabilityIam
CAPABILITY_IAM

OnFailure

data OnFailure

Constructors

OnFailureDelete
DELETE

OnFailureDoNothing
DO_NOTHING

OnFailureRollback
ROLLBACK

ResourceStatus

data ResourceStatus

Constructors

ResourceStatusCreateComplete
CREATE_COMPLETE

ResourceStatusCreateFailed
CREATE_FAILED

ResourceStatusCreateInProgress
CREATE_IN_PROGRESS

ResourceStatusDeleteComplete
DELETE_COMPLETE

ResourceStatusDeleteFailed
DELETE_FAILED

ResourceStatusDeleteInProgress
DELETE_IN_PROGRESS

ResourceStatusUpdateComplete
UPDATE_COMPLETE

ResourceStatusUpdateFailed
UPDATE_FAILED

ResourceStatusUpdateInProgress
UPDATE_IN_PROGRESS

StackStatus

data StackStatus

Constructors

StackStatusCreateComplete
CREATE_COMPLETE

StackStatusCreateFailed
CREATE_FAILED

StackStatusCreateInProgress
CREATE_IN_PROGRESS

StackStatusDeleteComplete
DELETE_COMPLETE

StackStatusDeleteFailed
DELETE_FAILED

StackStatusDeleteInProgress
DELETE_IN_PROGRESS

StackStatusRollbackComplete
ROLLBACK_COMPLETE

StackStatusRollbackFailed
ROLLBACK_FAILED

StackStatusRollbackInProgress
ROLLBACK_IN_PROGRESS

StackStatusUpdateComplete
UPDATE_COMPLETE

StackStatusUpdateCompleteCleanupInProgress
UPDATE_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateInProgress
UPDATE_IN_PROGRESS

StackStatusUpdateRollbackComplete
UPDATE_ROLLBACK_COMPLETE

StackStatusUpdateRollbackCompleteCleanupInProgress
UPDATE_ROLLBACK_COMPLETE_CLEANUP_IN_PROGRESS

StackStatusUpdateRollbackFailed
UPDATE_ROLLBACK_FAILED

StackStatusUpdateRollbackInProgress
UPDATE_ROLLBACK_IN_PROGRESS

Output

data Output

The Output data type.

mkOutput :: Output

Smart constructor for the minimum required fields to construct a valid Output data type to populate a request.

oOutputKey :: Lens' Output (Maybe Text)

The key associated with the output.

oOutputValue :: Lens' Output (Maybe Text)

The value associated with the output.

oDescription :: Lens' Output (Maybe Text)

User defined description associated with the output.

Parameter

data Parameter

The Parameter data type.

mkParameter :: Parameter

Smart constructor for the minimum required fields to construct a valid Parameter data type to populate a request.

pParameterKey :: Lens' Parameter (Maybe Text)

The key associated with the parameter.

pParameterValue :: Lens' Parameter (Maybe Text)

The value associated with the parameter.

pUsePreviousValue :: Lens' Parameter (Maybe Bool)

During a stack update, use the existing parameter value that is being used for the stack.

Stack

data Stack

The Stack data type.

mkStack

Arguments

:: Text
sStackName

-> ISO8601
sCreationTime

-> StackStatus
sStackStatus

-> Stack

Smart constructor for the minimum required fields to construct a valid Stack data type to populate a request.

sStackId :: Lens' Stack (Maybe Text)

Unique identifier of the stack.

sStackName :: Lens' Stack Text

The name associated with the stack.

sDescription :: Lens' Stack (Maybe Text)

User defined description associated with the stack.

sParameters :: Lens' Stack [Parameter]

A list of Parameter structures.

sCreationTime :: Lens' Stack ISO8601

Time at which the stack was created.

sLastUpdatedTime :: Lens' Stack (Maybe ISO8601)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

sStackStatus :: Lens' Stack StackStatus

Current status of the stack.

sStackStatusReason :: Lens' Stack (Maybe Text)

Success/failure message associated with the stack status.

sDisableRollback :: Lens' Stack (Maybe Bool)

Boolean to enable or disable rollback on stack creation failures: true: disable rollback false: enable rollback.

sNotificationARNs :: Lens' Stack [Text]

SNS topic ARNs to which stack related events are published.

sTimeoutInMinutes :: Lens' Stack (Maybe Integer)

The amount of time within which stack creation should complete.

sCapabilities :: Lens' Stack [Capability]

The capabilities allowed in the stack.

sOutputs :: Lens' Stack [Output]

A list of output structures.

sTags :: Lens' Stack [Tag]

A list of Tags that specify cost allocation information for the stack.

StackEvent

data StackEvent

The StackEvent data type.

mkStackEvent

Arguments

:: Text
seStackId

-> Text
seEventId

-> Text
seStackName

-> ISO8601
seTimestamp

-> StackEvent

Smart constructor for the minimum required fields to construct a valid StackEvent data type to populate a request.

seStackId :: Lens' StackEvent Text

The unique ID name of the instance of the stack.

seEventId :: Lens' StackEvent Text

The unique ID of this event.

seStackName :: Lens' StackEvent Text

The name associated with a stack.

seLogicalResourceId :: Lens' StackEvent (Maybe Text)

The logical name of the resource specified in the template.

sePhysicalResourceId :: Lens' StackEvent (Maybe Text)

The name or unique identifier associated with the physical instance of the resource.

seResourceType :: Lens' StackEvent (Maybe Text)

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

seTimestamp :: Lens' StackEvent ISO8601

Time the status was updated.

seResourceStatus :: Lens' StackEvent (Maybe ResourceStatus)

Current status of the resource.

seResourceStatusReason :: Lens' StackEvent (Maybe Text)

Success/failure message associated with the resource.

seResourceProperties :: Lens' StackEvent (Maybe Text)

BLOB of the properties used to create the resource.

StackResource

data StackResource

The StackResource data type.

mkStackResource

Arguments

:: Text
srLogicalResourceId

-> Text
srResourceType

-> ISO8601
srTimestamp

-> ResourceStatus
srResourceStatus

-> StackResource

Smart constructor for the minimum required fields to construct a valid StackResource data type to populate a request.

srStackName :: Lens' StackResource (Maybe Text)

The name associated with the stack.

srStackId :: Lens' StackResource (Maybe Text)

Unique identifier of the stack.

srLogicalResourceId :: Lens' StackResource Text

The logical name of the resource specified in the template.

srPhysicalResourceId :: Lens' StackResource (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

srResourceType :: Lens' StackResource Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

srTimestamp :: Lens' StackResource ISO8601

Time the status was updated.

srResourceStatus :: Lens' StackResource ResourceStatus

Current status of the resource.

srResourceStatusReason :: Lens' StackResource (Maybe Text)

Success/failure message associated with the resource.

srDescription :: Lens' StackResource (Maybe Text)

User defined description associated with the resource.

StackResourceDetail

data StackResourceDetail

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

mkStackResourceDetail

Arguments

:: Text
srdLogicalResourceId

-> Text
srdResourceType

-> ISO8601
srdLastUpdatedTimestamp

-> ResourceStatus
srdResourceStatus

-> StackResourceDetail

Smart constructor for the minimum required fields to construct a valid StackResourceDetail data type to populate a request.

srdStackName :: Lens' StackResourceDetail (Maybe Text)

The name associated with the stack.

srdStackId :: Lens' StackResourceDetail (Maybe Text)

Unique identifier of the stack.

srdLogicalResourceId :: Lens' StackResourceDetail Text

The logical name of the resource specified in the template.

srdPhysicalResourceId :: Lens' StackResourceDetail (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

srdResourceType :: Lens' StackResourceDetail Text

Type of resource. ((For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

srdLastUpdatedTimestamp :: Lens' StackResourceDetail ISO8601

Time the status was updated.

srdResourceStatus :: Lens' StackResourceDetail ResourceStatus

Current status of the resource.

srdResourceStatusReason :: Lens' StackResourceDetail (Maybe Text)

Success/failure message associated with the resource.

srdDescription :: Lens' StackResourceDetail (Maybe Text)

User defined description associated with the resource.

srdMetadata :: Lens' StackResourceDetail (Maybe Text)

The JSON format content of the Metadata attribute declared for the resource. For more information, see Metadata Attribute in the AWS CloudFormation User Guide.

StackResourceSummary

data StackResourceSummary

Contains high-level information about the specified stack resource.

mkStackResourceSummary

Arguments

:: Text
srsLogicalResourceId

-> Text
srsResourceType

-> ISO8601
srsLastUpdatedTimestamp

-> ResourceStatus
srsResourceStatus

-> StackResourceSummary

Smart constructor for the minimum required fields to construct a valid StackResourceSummary data type to populate a request.

srsLogicalResourceId :: Lens' StackResourceSummary Text

The logical name of the resource specified in the template.

srsPhysicalResourceId :: Lens' StackResourceSummary (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of the resource.

srsResourceType :: Lens' StackResourceSummary Text

Type of resource. (For more information, go to AWS Resource Types Reference in the AWS CloudFormation User Guide.).

srsLastUpdatedTimestamp :: Lens' StackResourceSummary ISO8601

Time the status was updated.

srsResourceStatus :: Lens' StackResourceSummary ResourceStatus

Current status of the resource.

srsResourceStatusReason :: Lens' StackResourceSummary (Maybe Text)

Success/failure message associated with the resource.

StackSummary

data StackSummary

The StackSummary Data Type.

mkStackSummary

Arguments

:: Text
ssStackName

-> ISO8601
ssCreationTime

-> StackStatus
ssStackStatus

-> StackSummary

Smart constructor for the minimum required fields to construct a valid StackSummary data type to populate a request.

ssStackId :: Lens' StackSummary (Maybe Text)

Unique stack identifier.

ssStackName :: Lens' StackSummary Text

The name associated with the stack.

ssTemplateDescription :: Lens' StackSummary (Maybe Text)

The template description of the template used to create the stack.

ssCreationTime :: Lens' StackSummary ISO8601

The time the stack was created.

ssLastUpdatedTime :: Lens' StackSummary (Maybe ISO8601)

The time the stack was last updated. This field will only be returned if the stack has been updated at least once.

ssDeletionTime :: Lens' StackSummary (Maybe ISO8601)

The time the stack was deleted.

ssStackStatus :: Lens' StackSummary StackStatus

The current status of the stack.

ssStackStatusReason :: Lens' StackSummary (Maybe Text)

Success/Failure message associated with the stack status.

Tag

data Tag

The Tag type is used by CreateStack in the Tags parameter. It allows you to specify a key/value pair that can be used to store information related to cost allocation for an AWS CloudFormation stack.

mkTag :: Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

tKey :: Lens' Tag (Maybe Text)

Required. A string used to identify this tag. You can specify a maximum of 128 characters for a tag key. Tags owned by Amazon Web Services (AWS) have the reserved prefix: aws:.

tValue :: Lens' Tag (Maybe Text)

Required. A string containing the value for this tag. You can specify a maximum of 256 characters for a tag value.

TemplateParameter

data TemplateParameter

The TemplateParameter data type.

mkTemplateParameter :: TemplateParameter

Smart constructor for the minimum required fields to construct a valid TemplateParameter data type to populate a request.

tpParameterKey :: Lens' TemplateParameter (Maybe Text)

The name associated with the parameter.

tpDefaultValue :: Lens' TemplateParameter (Maybe Text)

The default value associated with the parameter.

tpNoEcho :: Lens' TemplateParameter (Maybe Bool)

Flag indicating whether the parameter should be displayed as plain text in logs and UIs.

tpDescription :: Lens' TemplateParameter (Maybe Text)

User defined description associated with the parameter.