Network.AWS.CloudFormation.ValidateTemplate

Description

Validates a specified template.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_ValidateTemplate.html

Request

data ValidateTemplate

Request constructor

validateTemplate :: ValidateTemplate

ValidateTemplate constructor.

The fields accessible through corresponding lenses are:

vtTemplateBody :: Maybe Text

vtTemplateURL :: Maybe Text

Request lenses

vtTemplateBody :: Lens' ValidateTemplate (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

vtTemplateURL :: Lens' ValidateTemplate (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 307,200 bytes) located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

Response

data ValidateTemplateResponse

Response constructor

validateTemplateResponse :: ValidateTemplateResponse

ValidateTemplateResponse constructor.

The fields accessible through corresponding lenses are:

vtrCapabilities :: [Capability]

vtrCapabilitiesReason :: Maybe Text

vtrDescription :: Maybe Text

vtrParameters :: [TemplateParameter]

Response lenses

vtrCapabilities :: Lens' ValidateTemplateResponse [Capability]

The capabilities found within the template. Currently, AWS CloudFormation supports only the CAPABILITY_IAM capability. If your template contains IAM resources, you must specify the CAPABILITY_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error.

vtrCapabilitiesReason :: Lens' ValidateTemplateResponse (Maybe Text)

The capabilities reason found within the template.

vtrDescription :: Lens' ValidateTemplateResponse (Maybe Text)

The description found within the template.

vtrParameters :: Lens' ValidateTemplateResponse [TemplateParameter]

A list of TemplateParameter structures.