Network.AWS.CloudFormation.CancelUpdateStack

Description

Cancels an update on the specified stack. If the call completes successfully, the stack rolls back the update and reverts to the previous stack configuration.

Creating a Request

cancelUpdateStack

Arguments

:: Text
cusStackName

-> CancelUpdateStack

Creates a value of CancelUpdateStack with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cusClientRequestToken - A unique identifier for this CancelUpdateStack request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to cancel an update on a stack with the same name. You might retry CancelUpdateStack requests to ensure that AWS CloudFormation successfully received them.

cusStackName - The name or the unique stack ID that is associated with the stack.

data CancelUpdateStack

The input for the CancelUpdateStack action.

See: cancelUpdateStack smart constructor.

Request Lenses

cusClientRequestToken :: Lens' CancelUpdateStack (Maybe Text)

A unique identifier for this CancelUpdateStack request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to cancel an update on a stack with the same name. You might retry CancelUpdateStack requests to ensure that AWS CloudFormation successfully received them.

cusStackName :: Lens' CancelUpdateStack Text

The name or the unique stack ID that is associated with the stack.

Destructuring the Response

cancelUpdateStackResponse :: CancelUpdateStackResponse

Creates a value of CancelUpdateStackResponse with the minimum fields required to make a request.

data CancelUpdateStackResponse

See: cancelUpdateStackResponse smart constructor.