Network.AWS.CloudFormation.DescribeStackResource

Description

Returns a description of the specified resource in the specified stack.

For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted.

Creating a Request

describeStackResource

Arguments

:: Text
dsrsStackName

-> Text
dsrsLogicalResourceId

-> DescribeStackResource

Creates a value of DescribeStackResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsStackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsrsLogicalResourceId - The logical name of the resource as specified in the template. Default: There is no default value.

data DescribeStackResource

The input for DescribeStackResource action.

See: describeStackResource smart constructor.

Request Lenses

dsrsStackName :: Lens' DescribeStackResource Text

The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsrsLogicalResourceId :: Lens' DescribeStackResource Text

The logical name of the resource as specified in the template. Default: There is no default value.

Destructuring the Response

describeStackResourceResponse

Arguments

:: Int
dsrrsResponseStatus

-> DescribeStackResourceResponse

Creates a value of DescribeStackResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrrsStackResourceDetail - A StackResourceDetail structure containing the description of the specified resource in the specified stack.

dsrrsResponseStatus - -- | The response status code.

data DescribeStackResourceResponse

The output for a DescribeStackResource action.

See: describeStackResourceResponse smart constructor.

Response Lenses

dsrrsStackResourceDetail :: Lens' DescribeStackResourceResponse (Maybe StackResourceDetail)

A StackResourceDetail structure containing the description of the specified resource in the specified stack.

dsrrsResponseStatus :: Lens' DescribeStackResourceResponse Int

- | The response status code.