Network.AWS.CloudFormation.DescribeStackResources

Description

Returns AWS resource descriptions for running and deleted stacks. If StackName is specified, all the associated resources that are part of the stack are returned. If PhysicalResourceId is specified, the associated resources of the stack that the resource belongs to are returned.

For deleted stacks, DescribeStackResources returns resource information for up to 90 days after the stack has been deleted.

You must specify either StackName or PhysicalResourceId , but not both. In addition, you can specify LogicalResourceId to filter the returned result. For more information about resources, the LogicalResourceId and PhysicalResourceId , go to the AWS CloudFormation User Guide .

Creating a Request

describeStackResources :: DescribeStackResources

Creates a value of DescribeStackResources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrLogicalResourceId - The logical name of the resource as specified in the template. Default: There is no default value.

dsrPhysicalResourceId - The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation. For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId . You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack. Required: Conditional. If you do not specify PhysicalResourceId , you must specify StackName . Default: There is no default value.

dsrStackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value. Required: Conditional. If you do not specify StackName , you must specify PhysicalResourceId .

data DescribeStackResources

The input for DescribeStackResources action.

See: describeStackResources smart constructor.

Request Lenses

dsrLogicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The logical name of the resource as specified in the template. Default: There is no default value.

dsrPhysicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation. For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId . You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack. Required: Conditional. If you do not specify PhysicalResourceId , you must specify StackName . Default: There is no default value.

dsrStackName :: Lens' DescribeStackResources (Maybe Text)

The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value. Required: Conditional. If you do not specify StackName , you must specify PhysicalResourceId .

Destructuring the Response

describeStackResourcesResponse

Arguments

:: Int
dsrsrsResponseStatus

-> DescribeStackResourcesResponse

Creates a value of DescribeStackResourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsrsStackResources - A list of StackResource structures.

dsrsrsResponseStatus - -- | The response status code.

data DescribeStackResourcesResponse

The output for a DescribeStackResources action.

See: describeStackResourcesResponse smart constructor.

Response Lenses

dsrsrsStackResources :: Lens' DescribeStackResourcesResponse [StackResource]

A list of StackResource structures.

dsrsrsResponseStatus :: Lens' DescribeStackResourcesResponse Int

- | The response status code.