Network.AWS.CloudFormation.DescribeStackResources

Description

Returns AWS resource descriptions for running and deleted stacks. If StackName is specified, all the associated resources that are part of the stack are returned. If PhysicalResourceId is specified, the associated resources of the stack that the resource belongs to are returned.

Only the first 100 resources will be returned. If your stack has more resources than this, you should use ListStackResources instead. For deleted stacks, DescribeStackResources returns resource information for up to 90 days after the stack has been deleted.

You must specify either StackName or PhysicalResourceId, but not both. In addition, you can specify LogicalResourceId to filter the returned result. For more information about resources, the LogicalResourceId and PhysicalResourceId, go to the AWS CloudFormation User Guide.

A ValidationError is returned if you specify both StackName and PhysicalResourceId in the same request.

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_DescribeStackResources.html

Request

data DescribeStackResources

Request constructor

describeStackResources :: DescribeStackResources

DescribeStackResources constructor.

The fields accessible through corresponding lenses are:

dsrLogicalResourceId :: Maybe Text

dsrPhysicalResourceId :: Maybe Text

dsrStackName :: Maybe Text

Request lenses

dsrLogicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The logical name of the resource as specified in the template.

Default: There is no default value.

dsrPhysicalResourceId :: Lens' DescribeStackResources (Maybe Text)

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

For example, for an Amazon Elastic Compute Cloud (EC2) instance, PhysicalResourceId corresponds to the InstanceId. You can pass the EC2 InstanceId to DescribeStackResources to find which stack the instance belongs to and what other resources are part of the stack.

Required: Conditional. If you do not specify PhysicalResourceId, you must specify StackName.

Default: There is no default value.

dsrStackName :: Lens' DescribeStackResources (Maybe Text)

The name or the unique identifier associated with the stack, which are not always interchangeable:

Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Required: Conditional. If you do not specify StackName, you must specify PhysicalResourceId.

Response

data DescribeStackResourcesResponse

Response constructor

describeStackResourcesResponse :: DescribeStackResourcesResponse

DescribeStackResourcesResponse constructor.

The fields accessible through corresponding lenses are:

dsrrStackResources :: [StackResource]

Response lenses

dsrrStackResources :: Lens' DescribeStackResourcesResponse [StackResource]

A list of StackResource structures.