Network.AWS.CloudFormation.DescribeStacks

Description

Returns the description for the specified stack; if no stack name was specified, then it returns the description for all the stacks created.

This operation returns paginated results.

Creating a Request

describeStacks :: DescribeStacks

Creates a value of DescribeStacks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dNextToken - A string that identifies the next page of stacks that you want to retrieve.

dStackName - The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

data DescribeStacks

The input for DescribeStacks action.

See: describeStacks smart constructor.

Request Lenses

dNextToken :: Lens' DescribeStacks (Maybe Text)

A string that identifies the next page of stacks that you want to retrieve.

dStackName :: Lens' DescribeStacks (Maybe Text)

The name or the unique stack ID that is associated with the stack, which are not always interchangeable: * Running stacks: You can specify either the stack's name or its unique stack ID. * Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

Destructuring the Response

describeStacksResponse

Arguments

:: Int
dsrsResponseStatus

-> DescribeStacksResponse

Creates a value of DescribeStacksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsNextToken - If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

dsrsStacks - A list of stack structures.

dsrsResponseStatus - -- | The response status code.

data DescribeStacksResponse

The output for a DescribeStacks action.

See: describeStacksResponse smart constructor.

Response Lenses

dsrsNextToken :: Lens' DescribeStacksResponse (Maybe Text)

If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

dsrsStacks :: Lens' DescribeStacksResponse [Stack]

A list of stack structures.

dsrsResponseStatus :: Lens' DescribeStacksResponse Int

- | The response status code.