Network.AWS.CloudFormation.EstimateTemplateCost

Description

Returns the estimated monthly cost of a template. The return value is an AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

Creating a Request

estimateTemplateCost :: EstimateTemplateCost

Creates a value of EstimateTemplateCost with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etcParameters - A list of Parameter structures that specify input parameters.

etcTemplateBody - Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must pass TemplateBody or TemplateURL . If both are passed, only TemplateBody is used.

etcTemplateURL - Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody . If both are passed, only TemplateBody is used.

data EstimateTemplateCost

The input for an EstimateTemplateCost action.

See: estimateTemplateCost smart constructor.

Request Lenses

etcParameters :: Lens' EstimateTemplateCost [Parameter]

A list of Parameter structures that specify input parameters.

etcTemplateBody :: Lens' EstimateTemplateCost (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must pass TemplateBody or TemplateURL . If both are passed, only TemplateBody is used.

etcTemplateURL :: Lens' EstimateTemplateCost (Maybe Text)

Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody . If both are passed, only TemplateBody is used.

Destructuring the Response

estimateTemplateCostResponse

Arguments

:: Int
etcrsResponseStatus

-> EstimateTemplateCostResponse

Creates a value of EstimateTemplateCostResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

etcrsURL - An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

etcrsResponseStatus - -- | The response status code.

data EstimateTemplateCostResponse

The output for a EstimateTemplateCost action.

See: estimateTemplateCostResponse smart constructor.

Response Lenses

etcrsURL :: Lens' EstimateTemplateCostResponse (Maybe Text)

An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.

etcrsResponseStatus :: Lens' EstimateTemplateCostResponse Int

- | The response status code.