Network.AWS.CloudFormation.GetTemplateSummary

Description

Returns information about a new or existing template. The GetTemplateSummary action is useful for viewing parameter information, such as default parameter values and parameter types, before you create or update a stack.

You can use the GetTemplateSummary action when you submit a template, or you can get template information for a running or deleted stack.

For deleted stacks, GetTemplateSummary returns the template information for up to 90 days after the stack has been deleted. If the template does not exist, a ValidationError is returned.

Creating a Request

getTemplateSummary :: GetTemplateSummary

Creates a value of GetTemplateSummary with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsTemplateBody - Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

gtsTemplateURL - Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

gtsStackSetName - The name or unique ID of the stack set from which the stack was created.

gtsStackName - The name or the stack ID that is associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

data GetTemplateSummary

The input for the GetTemplateSummary action.

See: getTemplateSummary smart constructor.

Request Lenses

gtsTemplateBody :: Lens' GetTemplateSummary (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

gtsTemplateURL :: Lens' GetTemplateSummary (Maybe Text)

Location of file containing the template body. The URL must point to a template (max size: 460,800 bytes) that is located in an Amazon S3 bucket. For more information about templates, see Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

gtsStackSetName :: Lens' GetTemplateSummary (Maybe Text)

The name or unique ID of the stack set from which the stack was created.

gtsStackName :: Lens' GetTemplateSummary (Maybe Text)

The name or the stack ID that is associated with the stack, which are not always interchangeable. For running stacks, you can specify either the stack's name or its unique stack ID. For deleted stack, you must specify the unique stack ID. Conditional: You must specify only one of the following parameters: StackName , TemplateBody , or TemplateURL .

Destructuring the Response

getTemplateSummaryResponse

Arguments

:: Int
gtsrsResponseStatus

-> GetTemplateSummaryResponse

Creates a value of GetTemplateSummaryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtsrsDeclaredTransforms - A list of the transforms that are declared in the template.

gtsrsVersion - The AWS template format version, which identifies the capabilities of the template.

gtsrsCapabilitiesReason - The list of resources that generated the values in the Capabilities response element.

gtsrsParameters - A list of parameter declarations that describe various properties for each parameter.

gtsrsMetadata - The value that is defined for the Metadata property of the template.

gtsrsDescription - The value that is defined in the Description property of the template.

gtsrsCapabilities - The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates .

gtsrsResourceTypes - A list of all the template resource types that are defined in the template, such as AWS::EC2::Instance , AWS::Dynamo::Table , and Custom::MyCustomInstance .

gtsrsResponseStatus - -- | The response status code.

data GetTemplateSummaryResponse

The output for the GetTemplateSummary action.

See: getTemplateSummaryResponse smart constructor.

Response Lenses

gtsrsDeclaredTransforms :: Lens' GetTemplateSummaryResponse [Text]

A list of the transforms that are declared in the template.

gtsrsVersion :: Lens' GetTemplateSummaryResponse (Maybe Text)

The AWS template format version, which identifies the capabilities of the template.

gtsrsCapabilitiesReason :: Lens' GetTemplateSummaryResponse (Maybe Text)

The list of resources that generated the values in the Capabilities response element.

gtsrsParameters :: Lens' GetTemplateSummaryResponse [ParameterDeclaration]

A list of parameter declarations that describe various properties for each parameter.

gtsrsMetadata :: Lens' GetTemplateSummaryResponse (Maybe Text)

The value that is defined for the Metadata property of the template.

gtsrsDescription :: Lens' GetTemplateSummaryResponse (Maybe Text)

The value that is defined in the Description property of the template.

gtsrsCapabilities :: Lens' GetTemplateSummaryResponse [Capability]

The capabilities found within the template. If your template contains IAM resources, you must specify the CAPABILITY_IAM or CAPABILITY_NAMED_IAM value for this parameter when you use the CreateStack or UpdateStack actions with your template; otherwise, those actions return an InsufficientCapabilities error. For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates .

gtsrsResourceTypes :: Lens' GetTemplateSummaryResponse [Text]

A list of all the template resource types that are defined in the template, such as AWS::EC2::Instance , AWS::Dynamo::Table , and Custom::MyCustomInstance .

gtsrsResponseStatus :: Lens' GetTemplateSummaryResponse Int

- | The response status code.