Network.AWS.CloudFormation.ListStacks

Description

Returns the summary information for stacks whose status matches the specified StackStatusFilter. Summary information for stacks that have been deleted is kept for 90 days after the stack is deleted. If no StackStatusFilter is specified, summary information for all stacks is returned (including existing stacks and stacks that have been deleted).

This operation returns paginated results.

Creating a Request

listStacks :: ListStacks

Creates a value of ListStacks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsNextToken - A string that identifies the next page of stacks that you want to retrieve.

lsStackStatusFilter - Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

data ListStacks

The input for ListStacks action.

See: listStacks smart constructor.

Request Lenses

lsNextToken :: Lens' ListStacks (Maybe Text)

A string that identifies the next page of stacks that you want to retrieve.

lsStackStatusFilter :: Lens' ListStacks [StackStatus]

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

Destructuring the Response

listStacksResponse

Arguments

:: Int
lsrsResponseStatus

-> ListStacksResponse

Creates a value of ListStacksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsrsNextToken - If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

lsrsStackSummaries - A list of StackSummary structures containing information about the specified stacks.

lsrsResponseStatus - -- | The response status code.

data ListStacksResponse

The output for ListStacks action.

See: listStacksResponse smart constructor.

Response Lenses

lsrsNextToken :: Lens' ListStacksResponse (Maybe Text)

If the output exceeds 1 MB in size, a string that identifies the next page of stacks. If no additional page exists, this value is null.

lsrsStackSummaries :: Lens' ListStacksResponse [StackSummary]

A list of StackSummary structures containing information about the specified stacks.

lsrsResponseStatus :: Lens' ListStacksResponse Int

- | The response status code.