Network.AWS.CloudFormation.ListStacks

Description

Returns the summary information for stacks whose status matches the specified StackStatusFilter. Summary information for stacks that have been deleted is kept for 90 days after the stack is deleted. If no StackStatusFilter is specified, summary information for all stacks is returned (including existing stacks and stacks that have been deleted).

http://docs.aws.amazon.com/AWSCloudFormation/latest/APIReference/API_ListStacks.html

Request

data ListStacks

Request constructor

listStacks :: ListStacks

ListStacks constructor.

The fields accessible through corresponding lenses are:

lsNextToken :: Maybe Text

lsStackStatusFilter :: [StackStatus]

Request lenses

lsNextToken :: Lens' ListStacks (Maybe Text)

String that identifies the start of the next list of stacks, if there is one.

Default: There is no default value.

lsStackStatusFilter :: Lens' ListStacks [StackStatus]

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

Response

data ListStacksResponse

Response constructor

listStacksResponse :: ListStacksResponse

ListStacksResponse constructor.

The fields accessible through corresponding lenses are:

lsr1NextToken :: Maybe Text

lsr1StackSummaries :: [StackSummary]

Response lenses

lsr1NextToken :: Lens' ListStacksResponse (Maybe Text)

String that identifies the start of the next list of stacks, if there is one.

lsr1StackSummaries :: Lens' ListStacksResponse [StackSummary]

A list of StackSummary structures containing information about the specified stacks.