Network.AWS.CloudFormation.SetStackPolicy

Description

Sets a stack policy for a specified stack.

Creating a Request

setStackPolicy

Arguments

:: Text
sspStackName

-> SetStackPolicy

Creates a value of SetStackPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sspStackPolicyBody - Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackPolicyURL - Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackName - The name or unique stack ID that you want to associate a policy with.

data SetStackPolicy

The input for the SetStackPolicy action.

See: setStackPolicy smart constructor.

Request Lenses

sspStackPolicyBody :: Lens' SetStackPolicy (Maybe Text)

Structure containing the stack policy body. For more information, go to Prevent Updates to Stack Resources in the AWS CloudFormation User Guide. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackPolicyURL :: Lens' SetStackPolicy (Maybe Text)

Location of a file containing the stack policy. The URL must point to a policy (maximum size: 16 KB) located in an S3 bucket in the same region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

sspStackName :: Lens' SetStackPolicy Text

The name or unique stack ID that you want to associate a policy with.

Destructuring the Response

setStackPolicyResponse :: SetStackPolicyResponse

Creates a value of SetStackPolicyResponse with the minimum fields required to make a request.

data SetStackPolicyResponse

See: setStackPolicyResponse smart constructor.