Network.AWS.CloudFormation.SignalResource

Description

Sends a signal to the specified resource with a success or failure status. You can use the SignalResource API in conjunction with a creation policy or update policy. AWS CloudFormation doesn't proceed with a stack creation or update until resources receive the required number of signals or the timeout period is exceeded. The SignalResource API is useful in cases where you want to send signals from anywhere other than an Amazon EC2 instance.

Creating a Request

signalResource

Arguments

:: Text
sigStackName

-> Text
sigLogicalResourceId

-> Text
sigUniqueId

-> ResourceSignalStatus
sigStatus

-> SignalResource

Creates a value of SignalResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sigStackName - The stack name or unique stack ID that includes the resource that you want to signal.

sigLogicalResourceId - The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

sigUniqueId - A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

sigStatus - The status of the signal, which is either success or failure. A failure signal causes AWS CloudFormation to immediately fail the stack creation or update.

data SignalResource

The input for the SignalResource action.

See: signalResource smart constructor.

Request Lenses

sigStackName :: Lens' SignalResource Text

The stack name or unique stack ID that includes the resource that you want to signal.

sigLogicalResourceId :: Lens' SignalResource Text

The logical ID of the resource that you want to signal. The logical ID is the name of the resource that given in the template.

sigUniqueId :: Lens' SignalResource Text

A unique ID of the signal. When you signal Amazon EC2 instances or Auto Scaling groups, specify the instance ID that you are signaling as the unique ID. If you send multiple signals to a single resource (such as signaling a wait condition), each signal requires a different unique ID.

sigStatus :: Lens' SignalResource ResourceSignalStatus

The status of the signal, which is either success or failure. A failure signal causes AWS CloudFormation to immediately fail the stack creation or update.

Destructuring the Response

signalResourceResponse :: SignalResourceResponse

Creates a value of SignalResourceResponse with the minimum fields required to make a request.

data SignalResourceResponse

See: signalResourceResponse smart constructor.