Network.AWS.CloudFormation.V2010_05_15.DescribeStackEvents

Description

Returns all stack related events for a specified stack. For more information about a stack's event history, go to Stacks in the AWS CloudFormation User Guide. You can list events for stacks that have failed to create or have been deleted by specifying the unique stack identifier (stack ID). https:/cloudformation.us-east-1.amazonaws.com ?Action=DescribeStackEvents &StackName=MyStack &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] Event-1-Id arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 MyStack MyStack MyStack_One AWS::CloudFormation::Stack 2010-07-27T22:26:28Z CREATE_IN_PROGRESS User initiated Event-2-Id arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 MyStack MyDBInstance MyStack_DB1 AWS::SecurityGroup 2010-07-27T22:27:28Z CREATE_IN_PROGRESS {GroupDescription:...} Event-3-Id arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 MyStack MySG1 MyStack_SG1 AWS:: SecurityGroup 2010-07-27T22:28:28Z CREATE_COMPLETE.

Request

data DescribeStackEvents

The input for DescribeStackEvents action.

Request constructor

mkDescribeStackEvents

Arguments

:: Text
dseStackName

-> DescribeStackEvents

Smart constructor for the minimum required parameters to construct a valid DescribeStackEvents request.

The fields accessible through corresponding lenses are:

StackName :: Text

NextToken :: Maybe Text

Request lenses

dseStackName :: Lens' DescribeStackEvents Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dseNextToken :: Lens' DescribeStackEvents (Maybe Text)

String that identifies the start of the next list of events, if there is one. Default: There is no default value.

Response

data DescribeStackEventsResponse

The output for a DescribeStackEvents action.

Response constructor

mkDescribeStackEventsResponse :: DescribeStackEventsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeStackEventsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

StackEvents :: [StackEvent]

NextToken :: Maybe Text

Response lenses

dserStackEvents :: Lens' DescribeStackEventsResponse [StackEvent]

A list of StackEvents structures.

dserNextToken :: Lens' DescribeStackEventsResponse (Maybe Text)

String that identifies the start of the next list of events, if there is one.