Network.AWS.CloudFormation.V2010_05_15.DescribeStackResource

Description

Returns a description of the specified resource in the specified stack. For deleted stacks, DescribeStackResource returns resource information for up to 90 days after the stack has been deleted. https:/cloudformation.us-east-1.amazonaws.com ?Action=DescribeStackResource &StackName=MyStack &LogicalResourceId=MyDBInstance &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2011-07-08T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] arn:aws:cloudformation:us-east-1:123456789:stackMyStackaaf549a0-a413-11df-adb3-5081b3858e83 MyStack MyDBInstance MyStack_DB1 AWS::RDS::DBInstance 2011-07-07T22:27:28Z CREATE_COMPLETE.

Request

data DescribeStackResource

The input for DescribeStackResource action.

Request constructor

mkDescribeStackResource

Arguments

:: Text
dsrStackName

-> Text
dsrLogicalResourceId

-> DescribeStackResource

Smart constructor for the minimum required parameters to construct a valid DescribeStackResource request.

Request lenses

dsrStackName :: Lens' DescribeStackResource Text

The name or the unique identifier associated with the stack, which are not always interchangeable: Running stacks: You can specify either the stack's name or its unique stack ID. Deleted stacks: You must specify the unique stack ID. Default: There is no default value.

dsrLogicalResourceId :: Lens' DescribeStackResource Text

The logical name of the resource as specified in the template. Default: There is no default value.

Response

data DescribeStackResourceResponse

The output for a DescribeStackResource action.

Response lenses

dsrrsStackResourceDetail :: Lens' DescribeStackResourceResponse (Maybe StackResourceDetail)

A StackResourceDetail structure containing the description of the specified resource in the specified stack.