Network.AWS.CloudFormation.V2010_05_15.EstimateTemplateCost

Description

Returns the estimated monthly cost of a template. The return value is an AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template. https:/cloudformation.us-east-1.amazonaws.com ?Action=EstimateTemplateCost &TemplateURL= https:/s3.amazonaws.comcloudformation-samples-us-east-1/Drupal_Simple.template &Version=2010-05-15 &SignatureVersion=2 &Timestamp=2011-12-04T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] http:/calculator.s3.amazonaws.comcalc5.html?key=cf-2e351785-e821-450c-9d58-625e1e1ebfb6.

Request

data EstimateTemplateCost

Request constructor

mkEstimateTemplateCost :: EstimateTemplateCost

Smart constructor for the minimum required parameters to construct a valid EstimateTemplateCost request.

The fields accessible through corresponding lenses are:

TemplateBody :: Maybe Text

TemplateURL :: Maybe Text

Parameters :: [Parameter]

Request lenses

etcTemplateBody :: Lens' EstimateTemplateCost (Maybe Text)

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.) Conditional: You must pass TemplateBody or TemplateURL. If both are passed, only TemplateBody is used.

etcTemplateURL :: Lens' EstimateTemplateCost (Maybe Text)

Location of file containing the template body. The URL must point to a template located in an S3 bucket in the same region as the stack. For more information, go to Template Anatomy in the AWS CloudFormation User Guide. Conditional: You must pass TemplateURL or TemplateBody. If both are passed, only TemplateBody is used.

etcParameters :: Lens' EstimateTemplateCost [Parameter]

A list of Parameter structures that specify input parameters.

Response

data EstimateTemplateCostResponse

The output for a EstimateTemplateCost action.

Response constructor

mkEstimateTemplateCostResponse :: EstimateTemplateCostResponse

Smart constructor for the minimum required parameters to construct a valid EstimateTemplateCostResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Url :: Maybe Text

Response lenses

etcrUrl :: Lens' EstimateTemplateCostResponse (Maybe Text)

An AWS Simple Monthly Calculator URL with a query string that describes the resources required to run the template.