Network.AWS.CloudFormation.V2010_05_15.ListStacks

Description

Returns the summary information for stacks whose status matches the specified StackStatusFilter. Summary information for stacks that have been deleted is kept for 90 days after the stack is deleted. If no StackStatusFilter is specified, summary information for all stacks is returned (including existing stacks and stacks that have been deleted). https:/cloudformation.us-east-1.amazonaws.com ?Action=ListStacks &StackStatusFilter.member.1=CREATE_IN_PROGRESS &StackStatusFilter.member.2=DELETE_COMPLETE &Version=2010-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2010-07-27T22%3A26%3A28.000Z &AWSAccessKeyId=[AWS Access KeyID] &Signature=[Signature] arn:aws:cloudformation:us-east-1:1234567:stackTestCreate1aaaaa CREATE_IN_PROGRESS vpc1 2011-05-23T15:47:44Z Creates one EC2 instance and a load balancer. arn:aws:cloudformation:us-east-1:1234567:stackTestDelete2bbbbb DELETE_COMPLETE 2011-03-10T16:20:51Z WP1 2011-03-05T19:57:58Z A simple basic Cloudformation Template.

Request

data ListStacks

The input for ListStacks action.

Request constructor

mkListStacks :: ListStacks

Smart constructor for the minimum required parameters to construct a valid ListStacks request.

Request lenses

lsNextToken :: Lens' ListStacks (Maybe Text)

String that identifies the start of the next list of stacks, if there is one. Default: There is no default value.

lsStackStatusFilter :: Lens' ListStacks [StackStatus]

Stack status to use as a filter. Specify one or more stack status codes to list only stacks with the specified status codes. For a complete list of stack status codes, see the StackStatus parameter of the Stack data type.

Response

data ListStacksResponse

The output for ListStacks action.

Response constructor

mkListStacksResponse :: ListStacksResponse

Smart constructor for the minimum required parameters to construct a valid ListStacksResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

lsr1StackSummaries :: Lens' ListStacksResponse [StackSummary]

A list of StackSummary structures containing information about the specified stacks.

lsr1NextToken :: Lens' ListStacksResponse (Maybe Text)

String that identifies the start of the next list of stacks, if there is one.